@@ -4,7 +4,7 @@ version: 2.4.0-beta.1
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.8.0 <4.0.0"
 
 dependencies:
   archive: ^4.0.0
