@@ -1,6 +1,6 @@
 name: butterfly_tools
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.8.0 <4.0.0"
 dependencies:
   args: ^2.5.0
   intl: ^0.20.0
