@@ -7,30 +7,34 @@ import '../models/archive.dart';
 import '../models/data.dart';
 
 Archive convertTextDataToArchive(Map<String, dynamic> data) {
-  data = {
-    'fileVersion': 8,
-    ...legacyNoteDataJsonMigrator(data),
-  };
+  data = {'fileVersion': 8, ...legacyNoteDataJsonMigrator(data)};
   var reader = NoteData(Archive());
   reader = reader.setAsset(kMetaArchiveFile, utf8.encode(jsonEncode(data)));
   for (final palette in data['palettes'] ?? []) {
     reader = reader.setAsset(
-        '$kPalettesArchiveDirectory/${palette['name']}.json',
-        utf8.encode((jsonEncode(palette))));
+      '$kPalettesArchiveDirectory/${palette['name']}.json',
+      utf8.encode((jsonEncode(palette))),
+    );
   }
   for (final style in data['styles'] ?? []) {
     reader = reader.setAsset(
-        '$kStylesArchiveDirectory/${style.name}.json', utf8.encode((style)));
+      '$kStylesArchiveDirectory/${style.name}.json',
+      utf8.encode((style)),
+    );
   }
   for (final component in data['components'] ?? []) {
     reader = reader.setAsset(
-        '$kComponentsArchiveDirectory/${component.name}.json',
-        utf8.encode((component)));
+      '$kComponentsArchiveDirectory/${component.name}.json',
+      utf8.encode((component)),
+    );
   }
-  for (final page in (data['pages'] as Map<String, dynamic>?)?.entries ??
-      <MapEntry<String, dynamic>>[]) {
-    reader = reader.setAsset('$kPagesArchiveDirectory/${page.key}.json',
-        utf8.encode(jsonEncode(page.value)));
+  for (final page
+      in (data['pages'] as Map<String, dynamic>?)?.entries ??
+          <MapEntry<String, dynamic>>[]) {
+    reader = reader.setAsset(
+      '$kPagesArchiveDirectory/${page.key}.json',
+      utf8.encode(jsonEncode(page.value)),
+    );
   }
   final packData = data['pack'];
   final packs = packData is! Map<String, dynamic>
@@ -39,7 +43,8 @@ Archive convertTextDataToArchive(Map<String, dynamic> data) {
   for (final pack in packs?.entries ?? <MapEntry<String, NoteData>>[]) {
     reader = reader.setBundledPack(pack.value, pack.key);
   }
-  final assets = (data['assets'] as Map<String, dynamic>?)?.entries ??
+  final assets =
+      (data['assets'] as Map<String, dynamic>?)?.entries ??
       <MapEntry<String, dynamic>>[];
   for (final asset in assets) {
     reader = reader.setAsset(asset.key, base64Decode(asset.value));
@@ -115,8 +120,10 @@ Map<String, dynamic> convertDocumentToText(NoteData data) {
 
   final thumbnail = data.getThumbnail();
   if (thumbnail != null) {
-    output['thumbnail'] =
-        UriData.fromBytes(thumbnail, mimeType: 'image/png').toString();
+    output['thumbnail'] = UriData.fromBytes(
+      thumbnail,
+      mimeType: 'image/png',
+    ).toString();
   }
 
   return output;
@@ -141,31 +148,41 @@ Map<String, dynamic> legacyNoteDataJsonMigrator(Map<String, dynamic> data) {
 }
 
 Map<String, dynamic> _legacyDocumentJsonMigrator(
-    Map<String, dynamic> data, int? fileVersion) {
+  Map<String, dynamic> data,
+  int? fileVersion,
+) {
   if (fileVersion != null && fileVersion >= 0) {
     if (fileVersion < 4) {
       data['palettes'] = List.from(
-          Map<String, dynamic>.from(data['palettes'] ?? [])
-              .entries
-              .map((e) =>
-                  {'colors': List<int>.from(e.value).toList(), 'name': e.key})
-              .toList());
+        Map<String, dynamic>.from(data['palettes'] ?? []).entries
+            .map(
+              (e) => {
+                'colors': List<int>.from(e.value).toList(),
+                'name': e.key,
+              },
+            )
+            .toList(),
+      );
     }
     if (fileVersion < 5) {
-      data['painters'] = List.from((data['painters'] as List).map((e) {
-        var map = Map<String, dynamic>.from(e);
-        if (map['type'] == 'path-eraser') {
-          map['type'] = 'pathEraser';
-        }
-        return map;
-      })).where((map) => map['type'] != 'image').toList();
-      data['content'] = List.from((data['content'] as List).map((e) {
-        var map = Map<String, dynamic>.from(e);
-        if (map['type'] == 'paint') {
-          map['type'] = 'pen';
-        }
-        return map;
-      })).toList();
+      data['painters'] = List.from(
+        (data['painters'] as List).map((e) {
+          var map = Map<String, dynamic>.from(e);
+          if (map['type'] == 'path-eraser') {
+            map['type'] = 'pathEraser';
+          }
+          return map;
+        }),
+      ).where((map) => map['type'] != 'image').toList();
+      data['content'] = List.from(
+        (data['content'] as List).map((e) {
+          var map = Map<String, dynamic>.from(e);
+          if (map['type'] == 'paint') {
+            map['type'] = 'pen';
+          }
+          return map;
+        }),
+      ).toList();
       data['background'] = Map<String, dynamic>.from(data['background'] ?? {})
         ..['type'] = 'box';
     }
@@ -195,38 +212,43 @@ Map<String, dynamic> _legacyDocumentJsonMigrator(
       }
     }
     if (fileVersion < 7) {
-      data['content'] = List.from(data['content'] as List).where((e) {
-        return e['type'] != 'eraser';
-      }).map((e) {
-        if (e['type'] == 'label') {
-          return {
-            'type': 'text',
-            'property': {
-              'type': 'defined',
-              'paragraph': {
-                'type': 'defined',
-                'alignment': e['horizontalAlignment'],
-                'span': e,
-              },
-              'alignment': e['verticalAlignment'],
-            },
-          };
-        }
-        if (e['type'] == 'image') {
-          return {
-            ...e as Map,
-            'source': UriData.fromBytes(e['pixels'], mimeType: 'image/png'),
-          };
-        }
-        return e;
-      }).toList();
+      data['content'] = List.from(data['content'] as List)
+          .where((e) {
+            return e['type'] != 'eraser';
+          })
+          .map((e) {
+            if (e['type'] == 'label') {
+              return {
+                'type': 'text',
+                'property': {
+                  'type': 'defined',
+                  'paragraph': {
+                    'type': 'defined',
+                    'alignment': e['horizontalAlignment'],
+                    'span': e,
+                  },
+                  'alignment': e['verticalAlignment'],
+                },
+              };
+            }
+            if (e['type'] == 'image') {
+              return {
+                ...e as Map,
+                'source': UriData.fromBytes(e['pixels'], mimeType: 'image/png'),
+              };
+            }
+            return e;
+          })
+          .toList();
       if (data['createdAt'] is String) {
-        data['createdAt'] =
-            DateTime.parse(data['createdAt'] as String).millisecondsSinceEpoch;
+        data['createdAt'] = DateTime.parse(
+          data['createdAt'] as String,
+        ).millisecondsSinceEpoch;
       }
       if (data['updatedAt'] is String) {
-        data['updatedAt'] =
-            DateTime.parse(data['updatedAt'] as String).millisecondsSinceEpoch;
+        data['updatedAt'] = DateTime.parse(
+          data['updatedAt'] as String,
+        ).millisecondsSinceEpoch;
       }
     }
     if (fileVersion < 8) {
@@ -234,25 +256,25 @@ Map<String, dynamic> _legacyDocumentJsonMigrator(
         if (e['type'] == 'svg') {
           return {
             ...e,
-            'source': UriData.fromString(e['source'] as String,
-                encoding: utf8, mimeType: 'image/svg+xml'),
+            'source': UriData.fromString(
+              e['source'] as String,
+              encoding: utf8,
+              mimeType: 'image/svg+xml',
+            ),
           };
         }
         return e;
       }).toList();
       data['painters'] = List.from(data['painters'] as List).map((e) {
         if (e['type'] == 'stamp') {
-          return {
-            ...e,
-            'component': null,
-          };
+          return {...e, 'component': null};
         }
         return e;
       }).toList();
       data['pages'] = {
         'default': data['type'] == 'document'
             ? Map<String, dynamic>.from(data)
-            : data['document']
+            : data['document'],
       };
     }
   }
@@ -260,13 +282,19 @@ Map<String, dynamic> _legacyDocumentJsonMigrator(
 }
 
 Map<String, dynamic> _legacyTemplateJsonMigrator(
-    Map<String, dynamic> data, int? fileVersion) {
+  Map<String, dynamic> data,
+  int? fileVersion,
+) {
   data['document'] = _legacyDocumentJsonMigrator(
-      Map<String, dynamic>.from(data['document']), fileVersion);
+    Map<String, dynamic>.from(data['document']),
+    fileVersion,
+  );
   return {...data['document'], ...data};
 }
 
 Map<String, dynamic> _legacyPackJsonMigrator(
-    Map<String, dynamic> data, int? fileVersion) {
+  Map<String, dynamic> data,
+  int? fileVersion,
+) {
   return data;
 }
