@@ -30,41 +30,45 @@ String _exportColor(SRGBColor value) {
 }
 
 (NoteData, PadElement?) getElement(
-    NoteData data, XmlElement element, String collectionName) {
+  NoteData data,
+  XmlElement element,
+  String collectionName,
+) {
   PadElement? get() {
     switch (element.qualifiedName) {
       case 'stroke':
         return PenElement(
           property: PenProperty(
             color: _importColor(element.getAttribute('color')!),
-            strokeWidth:
-                double.parse(element.getAttribute('width')!.split(' ').first),
+            strokeWidth: double.parse(
+              element.getAttribute('width')!.split(' ').first,
+            ),
           ),
-          extra: {
-            'xopp:width': element.getAttribute('width')!,
-          },
-          points: toPoints(element.innerText
-              .split(' ')
-              .map((e) => double.parse(e))
-              .toList()),
+          extra: {'xopp:width': element.getAttribute('width')!},
+          points: toPoints(
+            element.innerText.split(' ').map((e) => double.parse(e)).toList(),
+          ),
           collection: collectionName,
         );
       case 'text':
         return TextElement(
           area: text.TextArea(
-              paragraph: text.TextParagraph(
-            textSpans: [
-              text.TextSpan.text(
-                text: element.innerText,
-                property: text.SpanProperty.defined(
-                  color: _importColor(element.getAttribute('color')!),
-                  size: double.parse(element.getAttribute('size')!),
+            paragraph: text.TextParagraph(
+              textSpans: [
+                text.TextSpan.text(
+                  text: element.innerText,
+                  property: text.SpanProperty.defined(
+                    color: _importColor(element.getAttribute('color')!),
+                    size: double.parse(element.getAttribute('size')!),
+                  ),
                 ),
-              ),
-            ],
-          )),
-          position: Point(double.parse(element.getAttribute('x')!),
-              double.parse(element.getAttribute('y')!)),
+              ],
+            ),
+          ),
+          position: Point(
+            double.parse(element.getAttribute('x')!),
+            double.parse(element.getAttribute('y')!),
+          ),
           collection: collectionName,
         );
       case 'image':
@@ -94,10 +98,12 @@ NoteData xoppMigrator(Uint8List data) {
   final doc = XmlDocument.parse(utf8.decode(GZipDecoder().decodeBytes(data)));
   final xournal = doc.getElement('xournal')!;
   var note = NoteData(Archive());
-  note = note.setMetadata(FileMetadata(
-    type: NoteFileType.document,
-    name: xournal.getElement('title')!.innerText,
-  ));
+  note = note.setMetadata(
+    FileMetadata(
+      type: NoteFileType.document,
+      name: xournal.getElement('title')!.innerText,
+    ),
+  );
   for (final entry in xournal.findElements('page').toList().asMap().entries) {
     final elements = <PadElement>[];
     final page = entry.value;
@@ -119,30 +125,29 @@ NoteData xoppMigrator(Uint8List data) {
     }
     final backgroundXml = page.getElement('background')!;
     final backgroundStyle = backgroundXml.getAttribute('style');
-    final backgroundColor =
-        _importColor(backgroundXml.getAttribute('color')!.substring(1));
+    final backgroundColor = _importColor(
+      backgroundXml.getAttribute('color')!.substring(1),
+    );
     final background = switch (backgroundXml.getAttribute('type')) {
       'solid' => Background.texture(
-            texture: SurfaceTexture.pattern(
+        texture: SurfaceTexture.pattern(
           boxXColor: backgroundColor,
           boxYColor: backgroundColor,
-          boxYSpace:
-              backgroundStyle == 'ruled' || backgroundStyle == 'lined' ? 20 : 0,
+          boxYSpace: backgroundStyle == 'ruled' || backgroundStyle == 'lined'
+              ? 20
+              : 0,
           boxXSpace: backgroundStyle == 'ruled' ? 20 : 0,
-        )),
+        ),
+      ),
       _ => null,
     };
-    (note, _) = note.addPage(DocumentPage(
-      layers: [
-        DocumentLayer(content: elements, id: createUniqueId()),
-      ],
-      backgrounds: [
-        if (background != null) background,
-      ],
-      extra: {
-        'xopp:layers': layers,
-      },
-    ));
+    (note, _) = note.addPage(
+      DocumentPage(
+        layers: [DocumentLayer(content: elements, id: createUniqueId())],
+        backgrounds: [if (background != null) background],
+        extra: {'xopp:layers': layers},
+      ),
+    );
   }
   return note;
 }
@@ -150,67 +155,106 @@ NoteData xoppMigrator(Uint8List data) {
 Uint8List xoppExporter(NoteData document) {
   final builder = XmlBuilder();
   builder.processing('xml', 'version="1.0" encoding="UTF-8"');
-  builder.element('xournal',
-      attributes: {'creator': 'Butterfly', 'fileversion': '4'}, nest: () {
-    final metadata = document.getMetadata();
-    builder.element('title', nest: metadata?.name);
-    for (final pageName in document.getPages()) {
-      final page = document.getPage(pageName);
-      if (page == null) continue;
-      builder.element('page', nest: () {
-        builder.element('background', attributes: {
-          'type': 'solid',
-          'color': _exportColor(
-              page.backgrounds.firstOrNull?.defaultColor ?? SRGBColor.white),
-          'style': 'plain',
-        });
-        builder.element('layer', nest: () {
-          for (final element in page.content) {
-            switch (element) {
-              case PenElement e:
-                builder.element('stroke', attributes: {
-                  'color': _exportColor(e.property.color),
-                  'width': e.property.strokeWidth.toString(),
-                  'tool': 'pen',
-                }, nest: () {
-                  builder.text(e.points.map((e) => '${e.x} ${e.y}').join(' '));
-                });
-                break;
-              case LabelElement e:
-                final styleSheet = e.styleSheet;
-                final style = e is TextElement
-                    ? styleSheet?.item
-                        .resolveParagraphProperty(e.area.paragraph.property)
-                        ?.span
-                    : styleSheet?.item.getParagraphProperty('p')?.span;
-                builder.element('text', attributes: {
-                  'color': _exportColor(style?.color ?? SRGBColor.black),
-                  'size': (style?.size ?? 12).toString(),
-                  'x': e.position.x.toString(),
-                  'y': e.position.y.toString(),
-                }, nest: () {
-                  builder.text(e.text);
-                });
-              case ImageElement e:
-                final imageData = document.getAsset(Uri.parse(e.source).path);
-                builder.element('image', attributes: {
-                  'left': e.position.x.toString(),
-                  'top': e.position.y.toString(),
-                  'right': (e.position.x + e.width).toString(),
-                  'bottom': (e.position.y + e.height).toString(),
-                }, nest: () {
-                  builder.text(
-                      UriData.fromBytes(imageData ?? [], mimeType: 'image/png')
-                          .toString());
-                });
-              default:
-                break;
-            }
-          }
-        });
-      });
-    }
-  });
-  return Uint8List.fromList(GZipEncoder()
-      .encode(utf8.encode(builder.buildDocument().toXmlString(pretty: true))));
+  builder.element(
+    'xournal',
+    attributes: {'creator': 'Butterfly', 'fileversion': '4'},
+    nest: () {
+      final metadata = document.getMetadata();
+      builder.element('title', nest: metadata?.name);
+      for (final pageName in document.getPages()) {
+        final page = document.getPage(pageName);
+        if (page == null) continue;
+        builder.element(
+          'page',
+          nest: () {
+            builder.element(
+              'background',
+              attributes: {
+                'type': 'solid',
+                'color': _exportColor(
+                  page.backgrounds.firstOrNull?.defaultColor ?? SRGBColor.white,
+                ),
+                'style': 'plain',
+              },
+            );
+            builder.element(
+              'layer',
+              nest: () {
+                for (final element in page.content) {
+                  switch (element) {
+                    case PenElement e:
+                      builder.element(
+                        'stroke',
+                        attributes: {
+                          'color': _exportColor(e.property.color),
+                          'width': e.property.strokeWidth.toString(),
+                          'tool': 'pen',
+                        },
+                        nest: () {
+                          builder.text(
+                            e.points.map((e) => '${e.x} ${e.y}').join(' '),
+                          );
+                        },
+                      );
+                      break;
+                    case LabelElement e:
+                      final styleSheet = e.styleSheet;
+                      final style = e is TextElement
+                          ? styleSheet?.item
+                                .resolveParagraphProperty(
+                                  e.area.paragraph.property,
+                                )
+                                ?.span
+                          : styleSheet?.item.getParagraphProperty('p')?.span;
+                      builder.element(
+                        'text',
+                        attributes: {
+                          'color': _exportColor(
+                            style?.color ?? SRGBColor.black,
+                          ),
+                          'size': (style?.size ?? 12).toString(),
+                          'x': e.position.x.toString(),
+                          'y': e.position.y.toString(),
+                        },
+                        nest: () {
+                          builder.text(e.text);
+                        },
+                      );
+                    case ImageElement e:
+                      final imageData = document.getAsset(
+                        Uri.parse(e.source).path,
+                      );
+                      builder.element(
+                        'image',
+                        attributes: {
+                          'left': e.position.x.toString(),
+                          'top': e.position.y.toString(),
+                          'right': (e.position.x + e.width).toString(),
+                          'bottom': (e.position.y + e.height).toString(),
+                        },
+                        nest: () {
+                          builder.text(
+                            UriData.fromBytes(
+                              imageData ?? [],
+                              mimeType: 'image/png',
+                            ).toString(),
+                          );
+                        },
+                      );
+                    default:
+                      break;
+                  }
+                }
+              },
+            );
+          },
+        );
+      }
+    },
+  );
+  return Uint8List.fromList(
+    GZipEncoder().encode(
+      utf8.encode(builder.buildDocument().toXmlString(pretty: true)),
+    ),
+  );
 }
