@@ -10,63 +10,48 @@ part 'search.freezed.dart';
 sealed class SearchResult with _$SearchResult {
   const SearchResult._();
 
-  const factory SearchResult.page(
-    String name,
-  ) = PageResult;
+  const factory SearchResult.page(String name) = PageResult;
   const factory SearchResult.element(
     String text,
     String page,
     Point<double> position,
     PadElement element,
   ) = ElementResult;
-  const factory SearchResult.area(
-    Area area,
-    String page,
-  ) = AreaResult;
-  const factory SearchResult.waypoint(
-    Waypoint waypoint,
-    String page,
-  ) = WaypointResult;
-  const factory SearchResult.tool(
-    String name,
-    Tool tool,
-    int index,
-  ) = ToolResult;
+  const factory SearchResult.area(Area area, String page) = AreaResult;
+  const factory SearchResult.waypoint(Waypoint waypoint, String page) =
+      WaypointResult;
+  const factory SearchResult.tool(String name, Tool tool, int index) =
+      ToolResult;
 
   String? getPage() => switch (this) {
-        PageResult e => e.name,
-        ElementResult e => e.page,
-        AreaResult e => e.page,
-        WaypointResult e => e.page,
-        _ => null,
-      };
+    PageResult e => e.name,
+    ElementResult e => e.page,
+    AreaResult e => e.page,
+    WaypointResult e => e.page,
+    _ => null,
+  };
 
   Point<double>? getPosition() => switch (this) {
-        ElementResult e => e.position,
-        AreaResult e => e.area.position,
-        WaypointResult e => e.waypoint.position,
-        _ => null,
-      };
+    ElementResult e => e.position,
+    AreaResult e => e.area.position,
+    WaypointResult e => e.waypoint.position,
+    _ => null,
+  };
 
   SearchResult withPage(String page) => switch (this) {
-        ElementResult e => e.copyWith(page: page),
-        PageResult e => e.copyWith(name: page),
-        AreaResult e => e.copyWith(page: page),
-        WaypointResult e => e.copyWith(page: page),
-        _ => this,
-      };
+    ElementResult e => e.copyWith(page: page),
+    PageResult e => e.copyWith(name: page),
+    AreaResult e => e.copyWith(page: page),
+    WaypointResult e => e.copyWith(page: page),
+    _ => this,
+  };
 }
 
 extension ElementSearchHelper on PadElement {
   ElementResult? matches(Pattern query) {
     ElementResult? matchText(TextElement e) {
       if (!e.area.paragraph.text.contains(query)) return null;
-      return ElementResult(
-        e.area.paragraph.text,
-        '',
-        e.position,
-        this,
-      );
+      return ElementResult(e.area.paragraph.text, '', e.position, this);
     }
 
     return switch (this) {
@@ -86,10 +71,7 @@ extension AreaSearchHelper on Area {
 extension WaypointSearchHelper on Waypoint {
   WaypointResult? matches(Pattern query) {
     if (name != null && name!.contains(query)) return null;
-    return WaypointResult(
-      this,
-      '',
-    );
+    return WaypointResult(this, '');
   }
 }
 
@@ -127,11 +109,15 @@ extension DocumentInfoSearchHelper on DocumentInfo {
 }
 
 extension GlobalSearchHelper on NoteData {
-  Iterable<SearchResult> search(Pattern query,
-      [String? currentPageName, DocumentPage? currentPage]) sync* {
+  Iterable<SearchResult> search(
+    Pattern query, [
+    String? currentPageName,
+    DocumentPage? currentPage,
+  ]) sync* {
     for (final pageName in getPages()) {
-      final page =
-          currentPageName == pageName ? currentPage : getPage(pageName);
+      final page = currentPageName == pageName
+          ? currentPage
+          : getPage(pageName);
       if (page == null) continue;
       if (pageName.contains(query)) yield PageResult(pageName);
       yield* page.search(query).map((e) => e.withPage(pageName));
