@@ -15,229 +15,152 @@ T _$identity<T>(T value) => value;
 
 /// @nodoc
 mixin _$Area {
-  String get name;
-  double get width;
-  double get height;
-  @DoublePointJsonConverter()
-  Point<double> get position;
-  Map<String, dynamic> get extra;
 
-  /// Create a copy of Area
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $AreaCopyWith<Area> get copyWith =>
-      _$AreaCopyWithImpl<Area>(this as Area, _$identity);
+ String get name; double get width; double get height;@DoublePointJsonConverter() Point<double> get position; Map<String, dynamic> get extra;
+/// Create a copy of Area
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$AreaCopyWith<Area> get copyWith => _$AreaCopyWithImpl<Area>(this as Area, _$identity);
 
   /// Serializes this Area to a JSON map.
   Map<String, dynamic> toJson();
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is Area &&
-            (identical(other.name, name) || other.name == name) &&
-            (identical(other.width, width) || other.width == width) &&
-            (identical(other.height, height) || other.height == height) &&
-            (identical(other.position, position) ||
-                other.position == position) &&
-            const DeepCollectionEquality().equals(other.extra, extra));
-  }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode => Object.hash(runtimeType, name, width, height, position,
-      const DeepCollectionEquality().hash(extra));
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is Area&&(identical(other.name, name) || other.name == name)&&(identical(other.width, width) || other.width == width)&&(identical(other.height, height) || other.height == height)&&(identical(other.position, position) || other.position == position)&&const DeepCollectionEquality().equals(other.extra, extra));
+}
 
-  @override
-  String toString() {
-    return 'Area(name: $name, width: $width, height: $height, position: $position, extra: $extra)';
-  }
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,name,width,height,position,const DeepCollectionEquality().hash(extra));
+
+@override
+String toString() {
+  return 'Area(name: $name, width: $width, height: $height, position: $position, extra: $extra)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $AreaCopyWith<$Res> {
-  factory $AreaCopyWith(Area value, $Res Function(Area) _then) =
-      _$AreaCopyWithImpl;
-  @useResult
-  $Res call(
-      {String name,
-      double width,
-      double height,
-      @DoublePointJsonConverter() Point<double> position,
-      Map<String, dynamic> extra});
-}
+abstract mixin class $AreaCopyWith<$Res>  {
+  factory $AreaCopyWith(Area value, $Res Function(Area) _then) = _$AreaCopyWithImpl;
+@useResult
+$Res call({
+ String name, double width, double height,@DoublePointJsonConverter() Point<double> position, Map<String, dynamic> extra
+});
 
+
+
+
+}
 /// @nodoc
-class _$AreaCopyWithImpl<$Res> implements $AreaCopyWith<$Res> {
+class _$AreaCopyWithImpl<$Res>
+    implements $AreaCopyWith<$Res> {
   _$AreaCopyWithImpl(this._self, this._then);
 
   final Area _self;
   final $Res Function(Area) _then;
 
-  /// Create a copy of Area
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? name = null,
-    Object? width = null,
-    Object? height = null,
-    Object? position = null,
-    Object? extra = null,
-  }) {
-    return _then(_self.copyWith(
-      name: null == name
-          ? _self.name
-          : name // ignore: cast_nullable_to_non_nullable
-              as String,
-      width: null == width
-          ? _self.width
-          : width // ignore: cast_nullable_to_non_nullable
-              as double,
-      height: null == height
-          ? _self.height
-          : height // ignore: cast_nullable_to_non_nullable
-              as double,
-      position: null == position
-          ? _self.position
-          : position // ignore: cast_nullable_to_non_nullable
-              as Point<double>,
-      extra: null == extra
-          ? _self.extra
-          : extra // ignore: cast_nullable_to_non_nullable
-              as Map<String, dynamic>,
-    ));
-  }
+/// Create a copy of Area
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? name = null,Object? width = null,Object? height = null,Object? position = null,Object? extra = null,}) {
+  return _then(_self.copyWith(
+name: null == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
+as String,width: null == width ? _self.width : width // ignore: cast_nullable_to_non_nullable
+as double,height: null == height ? _self.height : height // ignore: cast_nullable_to_non_nullable
+as double,position: null == position ? _self.position : position // ignore: cast_nullable_to_non_nullable
+as Point<double>,extra: null == extra ? _self.extra : extra // ignore: cast_nullable_to_non_nullable
+as Map<String, dynamic>,
+  ));
 }
+
+}
+
 
 /// @nodoc
 @JsonSerializable()
+
 class _Area extends Area {
-  const _Area(
-      {this.name = '',
-      required this.width,
-      required this.height,
-      @DoublePointJsonConverter() required this.position,
-      final Map<String, dynamic> extra = const {}})
-      : _extra = extra,
-        super._();
+  const _Area({this.name = '', required this.width, required this.height, @DoublePointJsonConverter() required this.position, final  Map<String, dynamic> extra = const {}}): _extra = extra,super._();
   factory _Area.fromJson(Map<String, dynamic> json) => _$AreaFromJson(json);
 
-  @override
-  @JsonKey()
-  final String name;
-  @override
-  final double width;
-  @override
-  final double height;
-  @override
-  @DoublePointJsonConverter()
-  final Point<double> position;
-  final Map<String, dynamic> _extra;
-  @override
-  @JsonKey()
-  Map<String, dynamic> get extra {
-    if (_extra is EqualUnmodifiableMapView) return _extra;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableMapView(_extra);
-  }
+@override@JsonKey() final  String name;
+@override final  double width;
+@override final  double height;
+@override@DoublePointJsonConverter() final  Point<double> position;
+ final  Map<String, dynamic> _extra;
+@override@JsonKey() Map<String, dynamic> get extra {
+  if (_extra is EqualUnmodifiableMapView) return _extra;
+  // ignore: implicit_dynamic_type
+  return EqualUnmodifiableMapView(_extra);
+}
 
-  /// Create a copy of Area
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$AreaCopyWith<_Area> get copyWith =>
-      __$AreaCopyWithImpl<_Area>(this, _$identity);
 
-  @override
-  Map<String, dynamic> toJson() {
-    return _$AreaToJson(
-      this,
-    );
-  }
+/// Create a copy of Area
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$AreaCopyWith<_Area> get copyWith => __$AreaCopyWithImpl<_Area>(this, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _Area &&
-            (identical(other.name, name) || other.name == name) &&
-            (identical(other.width, width) || other.width == width) &&
-            (identical(other.height, height) || other.height == height) &&
-            (identical(other.position, position) ||
-                other.position == position) &&
-            const DeepCollectionEquality().equals(other._extra, _extra));
-  }
+@override
+Map<String, dynamic> toJson() {
+  return _$AreaToJson(this, );
+}
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode => Object.hash(runtimeType, name, width, height, position,
-      const DeepCollectionEquality().hash(_extra));
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _Area&&(identical(other.name, name) || other.name == name)&&(identical(other.width, width) || other.width == width)&&(identical(other.height, height) || other.height == height)&&(identical(other.position, position) || other.position == position)&&const DeepCollectionEquality().equals(other._extra, _extra));
+}
 
-  @override
-  String toString() {
-    return 'Area(name: $name, width: $width, height: $height, position: $position, extra: $extra)';
-  }
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,name,width,height,position,const DeepCollectionEquality().hash(_extra));
+
+@override
+String toString() {
+  return 'Area(name: $name, width: $width, height: $height, position: $position, extra: $extra)';
+}
+
+
 }
 
 /// @nodoc
 abstract mixin class _$AreaCopyWith<$Res> implements $AreaCopyWith<$Res> {
-  factory _$AreaCopyWith(_Area value, $Res Function(_Area) _then) =
-      __$AreaCopyWithImpl;
-  @override
-  @useResult
-  $Res call(
-      {String name,
-      double width,
-      double height,
-      @DoublePointJsonConverter() Point<double> position,
-      Map<String, dynamic> extra});
-}
+  factory _$AreaCopyWith(_Area value, $Res Function(_Area) _then) = __$AreaCopyWithImpl;
+@override @useResult
+$Res call({
+ String name, double width, double height,@DoublePointJsonConverter() Point<double> position, Map<String, dynamic> extra
+});
 
+
+
+
+}
 /// @nodoc
-class __$AreaCopyWithImpl<$Res> implements _$AreaCopyWith<$Res> {
+class __$AreaCopyWithImpl<$Res>
+    implements _$AreaCopyWith<$Res> {
   __$AreaCopyWithImpl(this._self, this._then);
 
   final _Area _self;
   final $Res Function(_Area) _then;
 
-  /// Create a copy of Area
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? name = null,
-    Object? width = null,
-    Object? height = null,
-    Object? position = null,
-    Object? extra = null,
-  }) {
-    return _then(_Area(
-      name: null == name
-          ? _self.name
-          : name // ignore: cast_nullable_to_non_nullable
-              as String,
-      width: null == width
-          ? _self.width
-          : width // ignore: cast_nullable_to_non_nullable
-              as double,
-      height: null == height
-          ? _self.height
-          : height // ignore: cast_nullable_to_non_nullable
-              as double,
-      position: null == position
-          ? _self.position
-          : position // ignore: cast_nullable_to_non_nullable
-              as Point<double>,
-      extra: null == extra
-          ? _self._extra
-          : extra // ignore: cast_nullable_to_non_nullable
-              as Map<String, dynamic>,
-    ));
-  }
+/// Create a copy of Area
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? name = null,Object? width = null,Object? height = null,Object? position = null,Object? extra = null,}) {
+  return _then(_Area(
+name: null == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
+as String,width: null == width ? _self.width : width // ignore: cast_nullable_to_non_nullable
+as double,height: null == height ? _self.height : height // ignore: cast_nullable_to_non_nullable
+as double,position: null == position ? _self.position : position // ignore: cast_nullable_to_non_nullable
+as Point<double>,extra: null == extra ? _self._extra : extra // ignore: cast_nullable_to_non_nullable
+as Map<String, dynamic>,
+  ));
+}
+
+
 }
 
 // dart format on
