@@ -7,14 +7,14 @@ part of 'background.dart';
 // **************************************************************************
 
 TextureBackground _$TextureBackgroundFromJson(Map json) => TextureBackground(
-      texture: SurfaceTexture.fromJson(
-          Map<String, dynamic>.from(json['texture'] as Map)),
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  texture: SurfaceTexture.fromJson(
+    Map<String, dynamic>.from(json['texture'] as Map),
+  ),
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$TextureBackgroundToJson(TextureBackground instance) =>
     <String, dynamic>{
@@ -24,17 +24,16 @@ Map<String, dynamic> _$TextureBackgroundToJson(TextureBackground instance) =>
     };
 
 ImageBackground _$ImageBackgroundFromJson(Map json) => ImageBackground(
-      source: json['source'] as String,
-      width: (json['width'] as num).toDouble(),
-      height: (json['height'] as num).toDouble(),
-      scaleX: (json['scaleX'] as num?)?.toDouble() ?? 1,
-      scaleY: (json['scaleY'] as num?)?.toDouble() ?? 1,
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  source: json['source'] as String,
+  width: (json['width'] as num).toDouble(),
+  height: (json['height'] as num).toDouble(),
+  scaleX: (json['scaleX'] as num?)?.toDouble() ?? 1,
+  scaleY: (json['scaleY'] as num?)?.toDouble() ?? 1,
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ImageBackgroundToJson(ImageBackground instance) =>
     <String, dynamic>{
@@ -48,17 +47,16 @@ Map<String, dynamic> _$ImageBackgroundToJson(ImageBackground instance) =>
     };
 
 SvgBackground _$SvgBackgroundFromJson(Map json) => SvgBackground(
-      source: json['source'] as String,
-      width: (json['width'] as num).toDouble(),
-      height: (json['height'] as num).toDouble(),
-      scaleX: (json['scaleX'] as num?)?.toDouble() ?? 1,
-      scaleY: (json['scaleY'] as num?)?.toDouble() ?? 1,
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  source: json['source'] as String,
+  width: (json['width'] as num).toDouble(),
+  height: (json['height'] as num).toDouble(),
+  scaleX: (json['scaleX'] as num?)?.toDouble() ?? 1,
+  scaleY: (json['scaleY'] as num?)?.toDouble() ?? 1,
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$SvgBackgroundToJson(SvgBackground instance) =>
     <String, dynamic>{
