@@ -7,10 +7,10 @@ part of 'element.dart';
 // **************************************************************************
 
 _ElementConstraint _$ElementConstraintFromJson(Map json) => _ElementConstraint(
-      size: (json['size'] as num?)?.toDouble() ?? 0,
-      length: (json['length'] as num?)?.toDouble() ?? 0,
-      includeArea: json['includeArea'] as bool? ?? true,
-    );
+  size: (json['size'] as num?)?.toDouble() ?? 0,
+  length: (json['length'] as num?)?.toDouble() ?? 0,
+  includeArea: json['includeArea'] as bool? ?? true,
+);
 
 Map<String, dynamic> _$ElementConstraintToJson(_ElementConstraint instance) =>
     <String, dynamic>{
@@ -27,12 +27,12 @@ ScaledElementConstraints _$ScaledElementConstraintsFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$ScaledElementConstraintsToJson(
-        ScaledElementConstraints instance) =>
-    <String, dynamic>{
-      'scaleX': instance.scaleX,
-      'scaleY': instance.scaleY,
-      'type': instance.$type,
-    };
+  ScaledElementConstraints instance,
+) => <String, dynamic>{
+  'scaleX': instance.scaleX,
+  'scaleY': instance.scaleY,
+  'type': instance.$type,
+};
 
 FixedElementConstraints _$FixedElementConstraintsFromJson(Map json) =>
     FixedElementConstraints(
@@ -42,12 +42,12 @@ FixedElementConstraints _$FixedElementConstraintsFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$FixedElementConstraintsToJson(
-        FixedElementConstraints instance) =>
-    <String, dynamic>{
-      'height': instance.height,
-      'width': instance.width,
-      'type': instance.$type,
-    };
+  FixedElementConstraints instance,
+) => <String, dynamic>{
+  'height': instance.height,
+  'width': instance.width,
+  'type': instance.$type,
+};
 
 DynamicElementConstraints _$DynamicElementConstraintsFromJson(Map json) =>
     DynamicElementConstraints(
@@ -59,35 +59,35 @@ DynamicElementConstraints _$DynamicElementConstraintsFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$DynamicElementConstraintsToJson(
-        DynamicElementConstraints instance) =>
-    <String, dynamic>{
-      'height': instance.height,
-      'width': instance.width,
-      'aspectRatio': instance.aspectRatio,
-      'includeArea': instance.includeArea,
-      'type': instance.$type,
-    };
+  DynamicElementConstraints instance,
+) => <String, dynamic>{
+  'height': instance.height,
+  'width': instance.width,
+  'aspectRatio': instance.aspectRatio,
+  'includeArea': instance.includeArea,
+  'type': instance.$type,
+};
 
 PenElement _$PenElementFromJson(Map json) => PenElement(
-      rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
-      collection: json['collection'] as String? ?? '',
-      id: const IdJsonConverter().fromJson(json['id'] as String?),
-      zoom: (json['zoom'] as num?)?.toDouble(),
-      points: (json['points'] as List<dynamic>?)
-              ?.map((e) =>
-                  PathPoint.fromJson(Map<String, dynamic>.from(e as Map)))
-              .toList() ??
-          const [],
-      property: json['property'] == null
-          ? const PenProperty()
-          : PenProperty.fromJson(
-              Map<String, dynamic>.from(json['property'] as Map)),
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
+  collection: json['collection'] as String? ?? '',
+  id: const IdJsonConverter().fromJson(json['id'] as String?),
+  zoom: (json['zoom'] as num?)?.toDouble(),
+  points:
+      (json['points'] as List<dynamic>?)
+          ?.map((e) => PathPoint.fromJson(Map<String, dynamic>.from(e as Map)))
+          .toList() ??
+      const [],
+  property: json['property'] == null
+      ? const PenProperty()
+      : PenProperty.fromJson(
+          Map<String, dynamic>.from(json['property'] as Map),
+        ),
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$PenElementToJson(PenElement instance) =>
     <String, dynamic>{
@@ -102,34 +102,36 @@ Map<String, dynamic> _$PenElementToJson(PenElement instance) =>
     };
 
 TextElement _$TextElementFromJson(Map json) => TextElement(
-      rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
-      collection: json['collection'] as String? ?? '',
-      id: const IdJsonConverter().fromJson(json['id'] as String?),
-      position: json['position'] == null
-          ? const Point(0.0, 0.0)
-          : const DoublePointJsonConverter().fromJson(json['position'] as Map),
-      scale: (json['scale'] as num?)?.toDouble() ?? 1.0,
-      styleSheet: json['styleSheet'] == null
-          ? null
-          : NamedItem<TextStyleSheet>.fromJson(
-              Map<String, dynamic>.from(json['styleSheet'] as Map),
-              (value) => TextStyleSheet.fromJson(
-                  Map<String, dynamic>.from(value as Map))),
-      area: TextArea.fromJson(Map<String, dynamic>.from(json['area'] as Map)),
-      constraint: json['constraint'] == null
-          ? const ElementConstraint(size: 1000)
-          : ElementConstraint.fromJson(
-              Map<String, dynamic>.from(json['constraint'] as Map)),
-      foreground: json['foreground'] == null
-          ? SRGBColor.black
-          : const ColorJsonConverter()
-              .fromJson((json['foreground'] as num).toInt()),
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
+  collection: json['collection'] as String? ?? '',
+  id: const IdJsonConverter().fromJson(json['id'] as String?),
+  position: json['position'] == null
+      ? const Point(0.0, 0.0)
+      : const DoublePointJsonConverter().fromJson(json['position'] as Map),
+  scale: (json['scale'] as num?)?.toDouble() ?? 1.0,
+  styleSheet: json['styleSheet'] == null
+      ? null
+      : NamedItem<TextStyleSheet>.fromJson(
+          Map<String, dynamic>.from(json['styleSheet'] as Map),
+          (value) =>
+              TextStyleSheet.fromJson(Map<String, dynamic>.from(value as Map)),
+        ),
+  area: TextArea.fromJson(Map<String, dynamic>.from(json['area'] as Map)),
+  constraint: json['constraint'] == null
+      ? const ElementConstraint(size: 1000)
+      : ElementConstraint.fromJson(
+          Map<String, dynamic>.from(json['constraint'] as Map),
+        ),
+  foreground: json['foreground'] == null
+      ? SRGBColor.black
+      : const ColorJsonConverter().fromJson(
+          (json['foreground'] as num).toInt(),
+        ),
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$TextElementToJson(TextElement instance) =>
     <String, dynamic>{
@@ -138,9 +140,7 @@ Map<String, dynamic> _$TextElementToJson(TextElement instance) =>
       'id': const IdJsonConverter().toJson(instance.id),
       'position': const DoublePointJsonConverter().toJson(instance.position),
       'scale': instance.scale,
-      'styleSheet': instance.styleSheet?.toJson(
-        (value) => value.toJson(),
-      ),
+      'styleSheet': instance.styleSheet?.toJson((value) => value.toJson()),
       'area': instance.area.toJson(),
       'constraint': instance.constraint.toJson(),
       'foreground': const ColorJsonConverter().toJson(instance.foreground),
@@ -149,38 +149,41 @@ Map<String, dynamic> _$TextElementToJson(TextElement instance) =>
     };
 
 MarkdownElement _$MarkdownElementFromJson(Map json) => MarkdownElement(
-      rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
-      collection: json['collection'] as String? ?? '',
-      id: const IdJsonConverter().fromJson(json['id'] as String?),
-      position: json['position'] == null
-          ? const Point(0.0, 0.0)
-          : const DoublePointJsonConverter().fromJson(json['position'] as Map),
-      scale: (json['scale'] as num?)?.toDouble() ?? 1.0,
-      styleSheet: json['styleSheet'] == null
-          ? null
-          : NamedItem<TextStyleSheet>.fromJson(
-              Map<String, dynamic>.from(json['styleSheet'] as Map),
-              (value) => TextStyleSheet.fromJson(
-                  Map<String, dynamic>.from(value as Map))),
-      areaProperty: json['areaProperty'] == null
-          ? const AreaProperty()
-          : AreaProperty.fromJson(
-              Map<String, dynamic>.from(json['areaProperty'] as Map)),
-      text: json['text'] as String,
-      constraint: json['constraint'] == null
-          ? const ElementConstraint(size: 1000)
-          : ElementConstraint.fromJson(
-              Map<String, dynamic>.from(json['constraint'] as Map)),
-      foreground: json['foreground'] == null
-          ? SRGBColor.black
-          : const ColorJsonConverter()
-              .fromJson((json['foreground'] as num).toInt()),
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
+  collection: json['collection'] as String? ?? '',
+  id: const IdJsonConverter().fromJson(json['id'] as String?),
+  position: json['position'] == null
+      ? const Point(0.0, 0.0)
+      : const DoublePointJsonConverter().fromJson(json['position'] as Map),
+  scale: (json['scale'] as num?)?.toDouble() ?? 1.0,
+  styleSheet: json['styleSheet'] == null
+      ? null
+      : NamedItem<TextStyleSheet>.fromJson(
+          Map<String, dynamic>.from(json['styleSheet'] as Map),
+          (value) =>
+              TextStyleSheet.fromJson(Map<String, dynamic>.from(value as Map)),
+        ),
+  areaProperty: json['areaProperty'] == null
+      ? const AreaProperty()
+      : AreaProperty.fromJson(
+          Map<String, dynamic>.from(json['areaProperty'] as Map),
+        ),
+  text: json['text'] as String,
+  constraint: json['constraint'] == null
+      ? const ElementConstraint(size: 1000)
+      : ElementConstraint.fromJson(
+          Map<String, dynamic>.from(json['constraint'] as Map),
+        ),
+  foreground: json['foreground'] == null
+      ? SRGBColor.black
+      : const ColorJsonConverter().fromJson(
+          (json['foreground'] as num).toInt(),
+        ),
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$MarkdownElementToJson(MarkdownElement instance) =>
     <String, dynamic>{
@@ -189,9 +192,7 @@ Map<String, dynamic> _$MarkdownElementToJson(MarkdownElement instance) =>
       'id': const IdJsonConverter().toJson(instance.id),
       'position': const DoublePointJsonConverter().toJson(instance.position),
       'scale': instance.scale,
-      'styleSheet': instance.styleSheet?.toJson(
-        (value) => value.toJson(),
-      ),
+      'styleSheet': instance.styleSheet?.toJson((value) => value.toJson()),
       'areaProperty': instance.areaProperty.toJson(),
       'text': instance.text,
       'constraint': instance.constraint.toJson(),
@@ -201,25 +202,25 @@ Map<String, dynamic> _$MarkdownElementToJson(MarkdownElement instance) =>
     };
 
 ImageElement _$ImageElementFromJson(Map json) => ImageElement(
-      rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
-      collection: json['collection'] as String? ?? '',
-      id: const IdJsonConverter().fromJson(json['id'] as String?),
-      position: json['position'] == null
-          ? const Point(0.0, 0.0)
-          : const DoublePointJsonConverter().fromJson(json['position'] as Map),
-      constraints: json['constraints'] == null
-          ? const ScaledElementConstraints(scaleX: 1, scaleY: 1)
-          : ElementConstraints.fromJson(
-              Map<String, dynamic>.from(json['constraints'] as Map)),
-      source: json['source'] as String,
-      width: (json['width'] as num).toDouble(),
-      height: (json['height'] as num).toDouble(),
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
+  collection: json['collection'] as String? ?? '',
+  id: const IdJsonConverter().fromJson(json['id'] as String?),
+  position: json['position'] == null
+      ? const Point(0.0, 0.0)
+      : const DoublePointJsonConverter().fromJson(json['position'] as Map),
+  constraints: json['constraints'] == null
+      ? const ScaledElementConstraints(scaleX: 1, scaleY: 1)
+      : ElementConstraints.fromJson(
+          Map<String, dynamic>.from(json['constraints'] as Map),
+        ),
+  source: json['source'] as String,
+  width: (json['width'] as num).toDouble(),
+  height: (json['height'] as num).toDouble(),
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ImageElementToJson(ImageElement instance) =>
     <String, dynamic>{
@@ -236,25 +237,25 @@ Map<String, dynamic> _$ImageElementToJson(ImageElement instance) =>
     };
 
 SvgElement _$SvgElementFromJson(Map json) => SvgElement(
-      rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
-      collection: json['collection'] as String? ?? '',
-      id: const IdJsonConverter().fromJson(json['id'] as String?),
-      position: json['position'] == null
-          ? const Point(0.0, 0.0)
-          : const DoublePointJsonConverter().fromJson(json['position'] as Map),
-      constraints: json['constraints'] == null
-          ? const ScaledElementConstraints(scaleX: 1, scaleY: 1)
-          : ElementConstraints.fromJson(
-              Map<String, dynamic>.from(json['constraints'] as Map)),
-      source: json['source'] as String,
-      width: (json['width'] as num).toDouble(),
-      height: (json['height'] as num).toDouble(),
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
+  collection: json['collection'] as String? ?? '',
+  id: const IdJsonConverter().fromJson(json['id'] as String?),
+  position: json['position'] == null
+      ? const Point(0.0, 0.0)
+      : const DoublePointJsonConverter().fromJson(json['position'] as Map),
+  constraints: json['constraints'] == null
+      ? const ScaledElementConstraints(scaleX: 1, scaleY: 1)
+      : ElementConstraints.fromJson(
+          Map<String, dynamic>.from(json['constraints'] as Map),
+        ),
+  source: json['source'] as String,
+  width: (json['width'] as num).toDouble(),
+  height: (json['height'] as num).toDouble(),
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$SvgElementToJson(SvgElement instance) =>
     <String, dynamic>{
@@ -271,64 +272,66 @@ Map<String, dynamic> _$SvgElementToJson(SvgElement instance) =>
     };
 
 ShapeElement _$ShapeElementFromJson(Map json) => ShapeElement(
-      rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
-      collection: json['collection'] as String? ?? '',
-      id: const IdJsonConverter().fromJson(json['id'] as String?),
-      firstPosition: json['firstPosition'] == null
-          ? const Point(0.0, 0.0)
-          : const DoublePointJsonConverter()
-              .fromJson(json['firstPosition'] as Map),
-      secondPosition: json['secondPosition'] == null
-          ? const Point(0.0, 0.0)
-          : const DoublePointJsonConverter()
-              .fromJson(json['secondPosition'] as Map),
-      property: json['property'] == null
-          ? const ShapeProperty(shape: RectangleShape())
-          : ShapeProperty.fromJson(
-              Map<String, dynamic>.from(json['property'] as Map)),
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
+  collection: json['collection'] as String? ?? '',
+  id: const IdJsonConverter().fromJson(json['id'] as String?),
+  firstPosition: json['firstPosition'] == null
+      ? const Point(0.0, 0.0)
+      : const DoublePointJsonConverter().fromJson(json['firstPosition'] as Map),
+  secondPosition: json['secondPosition'] == null
+      ? const Point(0.0, 0.0)
+      : const DoublePointJsonConverter().fromJson(
+          json['secondPosition'] as Map,
+        ),
+  property: json['property'] == null
+      ? const ShapeProperty(shape: RectangleShape())
+      : ShapeProperty.fromJson(
+          Map<String, dynamic>.from(json['property'] as Map),
+        ),
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ShapeElementToJson(ShapeElement instance) =>
     <String, dynamic>{
       'rotation': instance.rotation,
       'collection': instance.collection,
       'id': const IdJsonConverter().toJson(instance.id),
-      'firstPosition':
-          const DoublePointJsonConverter().toJson(instance.firstPosition),
-      'secondPosition':
-          const DoublePointJsonConverter().toJson(instance.secondPosition),
+      'firstPosition': const DoublePointJsonConverter().toJson(
+        instance.firstPosition,
+      ),
+      'secondPosition': const DoublePointJsonConverter().toJson(
+        instance.secondPosition,
+      ),
       'property': instance.property.toJson(),
       'extra': instance.extra,
       'type': instance.$type,
     };
 
 TextureElement _$TextureElementFromJson(Map json) => TextureElement(
-      rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
-      collection: json['collection'] as String? ?? '',
-      id: const IdJsonConverter().fromJson(json['id'] as String?),
-      texture: json['texture'] == null
-          ? const SurfaceTexture.pattern()
-          : SurfaceTexture.fromJson(
-              Map<String, dynamic>.from(json['texture'] as Map)),
-      firstPosition: json['firstPosition'] == null
-          ? const Point(0.0, 0.0)
-          : const DoublePointJsonConverter()
-              .fromJson(json['firstPosition'] as Map),
-      secondPosition: json['secondPosition'] == null
-          ? const Point(0.0, 0.0)
-          : const DoublePointJsonConverter()
-              .fromJson(json['secondPosition'] as Map),
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      $type: json['type'] as String?,
-    );
+  rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
+  collection: json['collection'] as String? ?? '',
+  id: const IdJsonConverter().fromJson(json['id'] as String?),
+  texture: json['texture'] == null
+      ? const SurfaceTexture.pattern()
+      : SurfaceTexture.fromJson(
+          Map<String, dynamic>.from(json['texture'] as Map),
+        ),
+  firstPosition: json['firstPosition'] == null
+      ? const Point(0.0, 0.0)
+      : const DoublePointJsonConverter().fromJson(json['firstPosition'] as Map),
+  secondPosition: json['secondPosition'] == null
+      ? const Point(0.0, 0.0)
+      : const DoublePointJsonConverter().fromJson(
+          json['secondPosition'] as Map,
+        ),
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$TextureElementToJson(TextureElement instance) =>
     <String, dynamic>{
@@ -336,33 +339,37 @@ Map<String, dynamic> _$TextureElementToJson(TextureElement instance) =>
       'collection': instance.collection,
       'id': const IdJsonConverter().toJson(instance.id),
       'texture': instance.texture.toJson(),
-      'firstPosition':
-          const DoublePointJsonConverter().toJson(instance.firstPosition),
-      'secondPosition':
-          const DoublePointJsonConverter().toJson(instance.secondPosition),
+      'firstPosition': const DoublePointJsonConverter().toJson(
+        instance.firstPosition,
+      ),
+      'secondPosition': const DoublePointJsonConverter().toJson(
+        instance.secondPosition,
+      ),
       'extra': instance.extra,
       'type': instance.$type,
     };
 
 PolygonElement _$PolygonElementFromJson(Map json) => PolygonElement(
-      rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
-      collection: json['collection'] as String? ?? '',
-      id: const IdJsonConverter().fromJson(json['id'] as String?),
-      points: (json['points'] as List<dynamic>?)
-              ?.map((e) =>
-                  PolygonPoint.fromJson(Map<String, dynamic>.from(e as Map)))
-              .toList() ??
-          const [],
-      extra: (json['extra'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          const {},
-      property: json['property'] == null
-          ? const PolygonProperty()
-          : PolygonProperty.fromJson(
-              Map<String, dynamic>.from(json['property'] as Map)),
-      $type: json['type'] as String?,
-    );
+  rotation: (json['rotation'] as num?)?.toDouble() ?? 0,
+  collection: json['collection'] as String? ?? '',
+  id: const IdJsonConverter().fromJson(json['id'] as String?),
+  points:
+      (json['points'] as List<dynamic>?)
+          ?.map(
+            (e) => PolygonPoint.fromJson(Map<String, dynamic>.from(e as Map)),
+          )
+          .toList() ??
+      const [],
+  extra:
+      (json['extra'] as Map?)?.map((k, e) => MapEntry(k as String, e)) ??
+      const {},
+  property: json['property'] == null
+      ? const PolygonProperty()
+      : PolygonProperty.fromJson(
+          Map<String, dynamic>.from(json['property'] as Map),
+        ),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$PolygonElementToJson(PolygonElement instance) =>
     <String, dynamic>{
