@@ -7,27 +7,35 @@ part of 'meta.dart';
 // **************************************************************************
 
 _FileMetadata _$FileMetadataFromJson(Map json) => _FileMetadata(
-      fileVersion: (json['fileVersion'] as num?)?.toInt(),
-      type: $enumDecode(_$NoteFileTypeEnumMap, json['type']),
-      createdAt: _$JsonConverterFromJson<int, DateTime>(
-          json['createdAt'], const DateTimeJsonConverter().fromJson),
-      updatedAt: _$JsonConverterFromJson<int, DateTime>(
-          json['updatedAt'], const DateTimeJsonConverter().fromJson),
-      name: json['name'] as String? ?? '',
-      description: json['description'] as String? ?? '',
-      author: json['author'] as String? ?? '',
-      directory: json['directory'] as String? ?? '',
-      version: json['version'] as String? ?? '',
-    );
+  fileVersion: (json['fileVersion'] as num?)?.toInt(),
+  type: $enumDecode(_$NoteFileTypeEnumMap, json['type']),
+  createdAt: _$JsonConverterFromJson<int, DateTime>(
+    json['createdAt'],
+    const DateTimeJsonConverter().fromJson,
+  ),
+  updatedAt: _$JsonConverterFromJson<int, DateTime>(
+    json['updatedAt'],
+    const DateTimeJsonConverter().fromJson,
+  ),
+  name: json['name'] as String? ?? '',
+  description: json['description'] as String? ?? '',
+  author: json['author'] as String? ?? '',
+  directory: json['directory'] as String? ?? '',
+  version: json['version'] as String? ?? '',
+);
 
 Map<String, dynamic> _$FileMetadataToJson(_FileMetadata instance) =>
     <String, dynamic>{
       'fileVersion': instance.fileVersion,
       'type': _$NoteFileTypeEnumMap[instance.type]!,
       'createdAt': _$JsonConverterToJson<int, DateTime>(
-          instance.createdAt, const DateTimeJsonConverter().toJson),
+        instance.createdAt,
+        const DateTimeJsonConverter().toJson,
+      ),
       'updatedAt': _$JsonConverterToJson<int, DateTime>(
-          instance.updatedAt, const DateTimeJsonConverter().toJson),
+        instance.updatedAt,
+        const DateTimeJsonConverter().toJson,
+      ),
       'name': instance.name,
       'description': instance.description,
       'author': instance.author,
@@ -44,11 +52,9 @@ const _$NoteFileTypeEnumMap = {
 Value? _$JsonConverterFromJson<Json, Value>(
   Object? json,
   Value? Function(Json json) fromJson,
-) =>
-    json == null ? null : fromJson(json as Json);
+) => json == null ? null : fromJson(json as Json);
 
 Json? _$JsonConverterToJson<Json, Value>(
   Value? value,
   Json? Function(Value value) toJson,
-) =>
-    value == null ? null : toJson(value);
+) => value == null ? null : toJson(value);
