@@ -7,12 +7,12 @@ part of 'pack.dart';
 // **************************************************************************
 
 _ColorPalette _$ColorPaletteFromJson(Map json) => _ColorPalette(
-      colors: (json['colors'] as List<dynamic>?)
-              ?.map((e) =>
-                  const ColorJsonConverter().fromJson((e as num).toInt()))
-              .toList() ??
-          const [],
-    );
+  colors:
+      (json['colors'] as List<dynamic>?)
+          ?.map((e) => const ColorJsonConverter().fromJson((e as num).toInt()))
+          .toList() ??
+      const [],
+);
 
 Map<String, dynamic> _$ColorPaletteToJson(_ColorPalette instance) =>
     <String, dynamic>{
@@ -22,10 +22,14 @@ Map<String, dynamic> _$ColorPaletteToJson(_ColorPalette instance) =>
 _ButterflyComponent _$ButterflyComponentFromJson(Map json) =>
     _ButterflyComponent(
       thumbnail: _$JsonConverterFromJson<String, Uint8List>(
-          json['thumbnail'], const Uint8ListJsonConverter().fromJson),
-      elements: (json['elements'] as List<dynamic>?)
-              ?.map((e) =>
-                  PadElement.fromJson(Map<String, dynamic>.from(e as Map)))
+        json['thumbnail'],
+        const Uint8ListJsonConverter().fromJson,
+      ),
+      elements:
+          (json['elements'] as List<dynamic>?)
+              ?.map(
+                (e) => PadElement.fromJson(Map<String, dynamic>.from(e as Map)),
+              )
               .toList() ??
           const <PadElement>[],
     );
@@ -33,28 +37,28 @@ _ButterflyComponent _$ButterflyComponentFromJson(Map json) =>
 Map<String, dynamic> _$ButterflyComponentToJson(_ButterflyComponent instance) =>
     <String, dynamic>{
       'thumbnail': _$JsonConverterToJson<String, Uint8List>(
-          instance.thumbnail, const Uint8ListJsonConverter().toJson),
+        instance.thumbnail,
+        const Uint8ListJsonConverter().toJson,
+      ),
       'elements': instance.elements.map((e) => e.toJson()).toList(),
     };
 
 Value? _$JsonConverterFromJson<Json, Value>(
   Object? json,
   Value? Function(Json json) fromJson,
-) =>
-    json == null ? null : fromJson(json as Json);
+) => json == null ? null : fromJson(json as Json);
 
 Json? _$JsonConverterToJson<Json, Value>(
   Value? value,
   Json? Function(Value value) toJson,
-) =>
-    value == null ? null : toJson(value);
+) => value == null ? null : toJson(value);
 
 TextParameter _$TextParameterFromJson(Map json) => TextParameter(
-      child: (json['child'] as num).toInt(),
-      name: json['name'] as String,
-      value: json['value'] as String,
-      $type: json['type'] as String?,
-    );
+  child: (json['child'] as num).toInt(),
+  name: json['name'] as String,
+  value: json['value'] as String,
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$TextParameterToJson(TextParameter instance) =>
     <String, dynamic>{
@@ -65,11 +69,11 @@ Map<String, dynamic> _$TextParameterToJson(TextParameter instance) =>
     };
 
 ColorParameter _$ColorParameterFromJson(Map json) => ColorParameter(
-      child: (json['child'] as num).toInt(),
-      name: json['name'] as String,
-      value: (json['value'] as num).toInt(),
-      $type: json['type'] as String?,
-    );
+  child: (json['child'] as num).toInt(),
+  name: json['name'] as String,
+  value: (json['value'] as num).toInt(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ColorParameterToJson(ColorParameter instance) =>
     <String, dynamic>{
@@ -80,11 +84,11 @@ Map<String, dynamic> _$ColorParameterToJson(ColorParameter instance) =>
     };
 
 BoolParameter _$BoolParameterFromJson(Map json) => BoolParameter(
-      child: (json['child'] as num).toInt(),
-      name: json['name'] as String,
-      value: json['value'] as bool,
-      $type: json['type'] as String?,
-    );
+  child: (json['child'] as num).toInt(),
+  name: json['name'] as String,
+  value: json['value'] as bool,
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$BoolParameterToJson(BoolParameter instance) =>
     <String, dynamic>{
@@ -95,11 +99,11 @@ Map<String, dynamic> _$BoolParameterToJson(BoolParameter instance) =>
     };
 
 IntParameter _$IntParameterFromJson(Map json) => IntParameter(
-      child: (json['child'] as num).toInt(),
-      name: json['name'] as String,
-      value: (json['value'] as num).toInt(),
-      $type: json['type'] as String?,
-    );
+  child: (json['child'] as num).toInt(),
+  name: json['name'] as String,
+  value: (json['value'] as num).toInt(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$IntParameterToJson(IntParameter instance) =>
     <String, dynamic>{
@@ -110,11 +114,11 @@ Map<String, dynamic> _$IntParameterToJson(IntParameter instance) =>
     };
 
 DoubleParameter _$DoubleParameterFromJson(Map json) => DoubleParameter(
-      child: (json['child'] as num).toInt(),
-      name: json['name'] as String,
-      value: (json['value'] as num).toDouble(),
-      $type: json['type'] as String?,
-    );
+  child: (json['child'] as num).toInt(),
+  name: json['name'] as String,
+  value: (json['value'] as num).toDouble(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$DoubleParameterToJson(DoubleParameter instance) =>
     <String, dynamic>{
@@ -127,17 +131,9 @@ Map<String, dynamic> _$DoubleParameterToJson(DoubleParameter instance) =>
 _NamedItem<T> _$NamedItemFromJson<T extends PackAsset>(
   Map json,
   T Function(Object? json) fromJsonT,
-) =>
-    _NamedItem<T>(
-      name: json['name'] as String,
-      item: fromJsonT(json['item']),
-    );
+) => _NamedItem<T>(name: json['name'] as String, item: fromJsonT(json['item']));
 
 Map<String, dynamic> _$NamedItemToJson<T extends PackAsset>(
   _NamedItem<T> instance,
   Object? Function(T value) toJsonT,
-) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'item': toJsonT(instance.item),
-    };
+) => <String, dynamic>{'name': instance.name, 'item': toJsonT(instance.item)};
