@@ -49,18 +49,16 @@ sealed class Property with _$Property {
 @freezed
 sealed class PathShape with _$PathShape {
   const PathShape._();
-  const factory PathShape.circle(
-      {@Default(SRGBColor.transparent)
-      @ColorJsonConverter()
-      SRGBColor fillColor}) = CircleShape;
-  const factory PathShape.rectangle(
-      {@Default(SRGBColor.transparent)
-      @ColorJsonConverter()
-      SRGBColor fillColor,
-      @Default(0) double topLeftCornerRadius,
-      @Default(0) double topRightCornerRadius,
-      @Default(0) double bottomLeftCornerRadius,
-      @Default(0) double bottomRightCornerRadius}) = RectangleShape;
+  const factory PathShape.circle({
+    @Default(SRGBColor.transparent) @ColorJsonConverter() SRGBColor fillColor,
+  }) = CircleShape;
+  const factory PathShape.rectangle({
+    @Default(SRGBColor.transparent) @ColorJsonConverter() SRGBColor fillColor,
+    @Default(0) double topLeftCornerRadius,
+    @Default(0) double topRightCornerRadius,
+    @Default(0) double bottomLeftCornerRadius,
+    @Default(0) double bottomRightCornerRadius,
+  }) = RectangleShape;
   const factory PathShape.line() = LineShape;
   const factory PathShape.triangle({
     @Default(SRGBColor.transparent) @ColorJsonConverter() SRGBColor fillColor,
