@@ -104,9 +104,7 @@ sealed class TextSpan with _$TextSpan {
     start = max(start, 0);
     length = length.clamp(0, this.length - start);
     final end = start + length;
-    return copyWith(
-      text: text.substring(start, end),
-    );
+    return copyWith(text: text.substring(start, end));
   }
 
   bool get isEmpty => text.isEmpty;
@@ -145,7 +143,9 @@ sealed class TextParagraph with _$TextParagraph {
     for (var span in textSpans) {
       if (currentLength + span.length > index) {
         return IndexedModel(
-            currentLength, cut ? span.subSpan(index - currentLength) : span);
+          currentLength,
+          cut ? span.subSpan(index - currentLength) : span,
+        );
       }
       currentLength += span.length;
     }
@@ -155,8 +155,11 @@ sealed class TextParagraph with _$TextParagraph {
   List<TextSpan> getSpans([int start = 0, int? length, bool cut = false]) =>
       getIndexedSpans(start, length, cut).model;
 
-  IndexedModel<List<TextSpan>> getIndexedSpans(
-      [int start = 0, int? length, bool cut = false]) {
+  IndexedModel<List<TextSpan>> getIndexedSpans([
+    int start = 0,
+    int? length,
+    bool cut = false,
+  ]) {
     length ??= this.length;
     var spans = <TextSpan>[];
     var currentLength = 0;
@@ -168,10 +171,14 @@ sealed class TextParagraph with _$TextParagraph {
           break;
         }
         firstIndex ??= currentLength;
-        spans.add(cut
-            ? span.subSpan(
-                start - currentLength, end - max(currentLength, start))
-            : span);
+        spans.add(
+          cut
+              ? span.subSpan(
+                  start - currentLength,
+                  end - max(currentLength, start),
+                )
+              : span,
+        );
       }
       currentLength += span.length;
     }
@@ -247,16 +254,21 @@ sealed class TextParagraph with _$TextParagraph {
     return copyWith(textSpans: subSpans);
   }
 
-  TextParagraph replaceText(String text,
-      [int start = 0, int? length, bool replaceWholeSpan = false]) {
+  TextParagraph replaceText(
+    String text, [
+    int start = 0,
+    int? length,
+    bool replaceWholeSpan = false,
+  ]) {
     final indexed = getIndexedSpan(start);
     length ??= replaceWholeSpan ? indexed?.model.length : 0;
     return replace(
-        indexed != null
-            ? indexed.model.copyWith(text: text)
-            : TextSpan.text(text: text),
-        (replaceWholeSpan ? indexed?.index : null) ?? start,
-        length);
+      indexed != null
+          ? indexed.model.copyWith(text: text)
+          : TextSpan.text(text: text),
+      (replaceWholeSpan ? indexed?.index : null) ?? start,
+      length,
+    );
   }
 
   TextParagraph remove([int start = 0, int length = 0]) {
@@ -264,11 +276,12 @@ sealed class TextParagraph with _$TextParagraph {
 
     final beforeSpans = getSpans(0, start, true);
     final afterSpans = getSpans(end, null, true);
-// Test if beforeSpans and afterSpans can be merged
+    // Test if beforeSpans and afterSpans can be merged
     if (beforeSpans.isNotEmpty &&
         beforeSpans.lastOrNull?.property == afterSpans.firstOrNull?.property) {
       final merged = beforeSpans.lastOrNull!.copyWith(
-          text: beforeSpans.lastOrNull!.text + afterSpans.firstOrNull!.text);
+        text: beforeSpans.lastOrNull!.text + afterSpans.firstOrNull!.text,
+      );
       afterSpans.removeAt(0);
       beforeSpans.removeLast();
       beforeSpans.add(merged);
@@ -276,20 +289,26 @@ sealed class TextParagraph with _$TextParagraph {
     return copyWith(textSpans: [...beforeSpans, ...afterSpans]);
   }
 
-  TextParagraph updateSpans(TextSpan Function(TextSpan) update,
-      [int start = 0, int length = 0]) {
+  TextParagraph updateSpans(
+    TextSpan Function(TextSpan) update, [
+    int start = 0,
+    int length = 0,
+  ]) {
     final spans = getSpans(start, length, true);
     if (spans.isEmpty) {
       return this;
     }
-    final updated = update(spans.first.copyWith(
-      text: spans.map((e) => e.text).join(),
-    ));
+    final updated = update(
+      spans.first.copyWith(text: spans.map((e) => e.text).join()),
+    );
     return replace(updated, start, length);
   }
 
-  TextParagraph applyStyle(SpanProperty property,
-      [int start = 0, int length = 0]) {
+  TextParagraph applyStyle(
+    SpanProperty property, [
+    int start = 0,
+    int length = 0,
+  ]) {
     return updateSpans(
       (span) => span.copyWith(property: property),
       start,
@@ -324,14 +343,7 @@ sealed class TextStyleSheet extends PackAsset with _$TextStyleSheet {
       _$TextStyleSheetFromJson(json);
 
   static List<String> getMarkdownSpanPropertyNames() {
-    return [
-      'em',
-      'strong',
-      'codeSpan',
-      'del',
-      'a',
-      'img',
-    ];
+    return ['em', 'strong', 'codeSpan', 'del', 'a', 'img'];
   }
 
   static List<String> getMarkdownParagraphPropertyNames() {
@@ -376,20 +388,21 @@ extension ResolveProperty on TextStyleSheet? {
       };
 
   DefinedParagraphProperty? resolveParagraphProperty(
-          ParagraphProperty? property) =>
-      switch (property) {
-        DefinedParagraphProperty e => e,
-        NamedParagraphProperty e => this?.paragraphProperties[e.name],
-        _ => null,
-      };
+    ParagraphProperty? property,
+  ) => switch (property) {
+    DefinedParagraphProperty e => e,
+    NamedParagraphProperty e => this?.paragraphProperties[e.name],
+    _ => null,
+  };
 }
 
 extension SpanPropertyGetter on DefinedSpanProperty {
   double getSize([DefinedParagraphProperty? paragraphProperty]) =>
       size ?? paragraphProperty?.span.size ?? SpanProperty.kDefault.size!;
-  SRGBColor getColor(
-          [DefinedParagraphProperty? paragraphProperty,
-          SRGBColor? foreground]) =>
+  SRGBColor getColor([
+    DefinedParagraphProperty? paragraphProperty,
+    SRGBColor? foreground,
+  ]) =>
       color ??
       paragraphProperty?.span.color ??
       foreground ??
@@ -420,13 +433,15 @@ extension SpanPropertyGetter on DefinedSpanProperty {
       decorationColor ??
       paragraphProperty?.span.decorationColor ??
       SpanProperty.kDefault.decorationColor!;
-  TextDecorationStyle getDecorationStyle(
-          [DefinedParagraphProperty? paragraphProperty]) =>
+  TextDecorationStyle getDecorationStyle([
+    DefinedParagraphProperty? paragraphProperty,
+  ]) =>
       decorationStyle ??
       paragraphProperty?.span.decorationStyle ??
       SpanProperty.kDefault.decorationStyle!;
-  double getDecorationThickness(
-          [DefinedParagraphProperty? paragraphProperty]) =>
+  double getDecorationThickness([
+    DefinedParagraphProperty? paragraphProperty,
+  ]) =>
       decorationThickness ??
       paragraphProperty?.span.decorationThickness ??
       SpanProperty.kDefault.decorationThickness!;
