@@ -10,7 +10,9 @@ DefinedSpanProperty _$DefinedSpanPropertyFromJson(Map json) =>
     DefinedSpanProperty(
       size: (json['size'] as num?)?.toDouble(),
       color: _$JsonConverterFromJson<int, SRGBColor>(
-          json['color'], const ColorJsonConverter().fromJson),
+        json['color'],
+        const ColorJsonConverter().fromJson,
+      ),
       fontWeight: (json['fontWeight'] as num?)?.toInt(),
       lineThrough: json['lineThrough'] as bool?,
       underline: json['underline'] as bool?,
@@ -18,41 +20,52 @@ DefinedSpanProperty _$DefinedSpanPropertyFromJson(Map json) =>
       italic: json['italic'] as bool?,
       letterSpacing: (json['letterSpacing'] as num?)?.toDouble(),
       decorationColor: _$JsonConverterFromJson<int, SRGBColor>(
-          json['decorationColor'], const ColorJsonConverter().fromJson),
+        json['decorationColor'],
+        const ColorJsonConverter().fromJson,
+      ),
       decorationStyle: $enumDecodeNullable(
-          _$TextDecorationStyleEnumMap, json['decorationStyle']),
+        _$TextDecorationStyleEnumMap,
+        json['decorationStyle'],
+      ),
       decorationThickness: (json['decorationThickness'] as num?)?.toDouble(),
       backgroundColor: _$JsonConverterFromJson<int, SRGBColor>(
-          json['backgroundColor'], const ColorJsonConverter().fromJson),
+        json['backgroundColor'],
+        const ColorJsonConverter().fromJson,
+      ),
       $type: json['type'] as String?,
     );
 
 Map<String, dynamic> _$DefinedSpanPropertyToJson(
-        DefinedSpanProperty instance) =>
-    <String, dynamic>{
-      'size': instance.size,
-      'color': _$JsonConverterToJson<int, SRGBColor>(
-          instance.color, const ColorJsonConverter().toJson),
-      'fontWeight': instance.fontWeight,
-      'lineThrough': instance.lineThrough,
-      'underline': instance.underline,
-      'overline': instance.overline,
-      'italic': instance.italic,
-      'letterSpacing': instance.letterSpacing,
-      'decorationColor': _$JsonConverterToJson<int, SRGBColor>(
-          instance.decorationColor, const ColorJsonConverter().toJson),
-      'decorationStyle': _$TextDecorationStyleEnumMap[instance.decorationStyle],
-      'decorationThickness': instance.decorationThickness,
-      'backgroundColor': _$JsonConverterToJson<int, SRGBColor>(
-          instance.backgroundColor, const ColorJsonConverter().toJson),
-      'type': instance.$type,
-    };
+  DefinedSpanProperty instance,
+) => <String, dynamic>{
+  'size': instance.size,
+  'color': _$JsonConverterToJson<int, SRGBColor>(
+    instance.color,
+    const ColorJsonConverter().toJson,
+  ),
+  'fontWeight': instance.fontWeight,
+  'lineThrough': instance.lineThrough,
+  'underline': instance.underline,
+  'overline': instance.overline,
+  'italic': instance.italic,
+  'letterSpacing': instance.letterSpacing,
+  'decorationColor': _$JsonConverterToJson<int, SRGBColor>(
+    instance.decorationColor,
+    const ColorJsonConverter().toJson,
+  ),
+  'decorationStyle': _$TextDecorationStyleEnumMap[instance.decorationStyle],
+  'decorationThickness': instance.decorationThickness,
+  'backgroundColor': _$JsonConverterToJson<int, SRGBColor>(
+    instance.backgroundColor,
+    const ColorJsonConverter().toJson,
+  ),
+  'type': instance.$type,
+};
 
 Value? _$JsonConverterFromJson<Json, Value>(
   Object? json,
   Value? Function(Json json) fromJson,
-) =>
-    json == null ? null : fromJson(json as Json);
+) => json == null ? null : fromJson(json as Json);
 
 const _$TextDecorationStyleEnumMap = {
   TextDecorationStyle.solid: 'solid',
@@ -65,50 +78,44 @@ const _$TextDecorationStyleEnumMap = {
 Json? _$JsonConverterToJson<Json, Value>(
   Value? value,
   Json? Function(Value value) toJson,
-) =>
-    value == null ? null : toJson(value);
+) => value == null ? null : toJson(value);
 
-NamedSpanProperty _$NamedSpanPropertyFromJson(Map json) => NamedSpanProperty(
-      json['name'] as String,
-      $type: json['type'] as String?,
-    );
+NamedSpanProperty _$NamedSpanPropertyFromJson(Map json) =>
+    NamedSpanProperty(json['name'] as String, $type: json['type'] as String?);
 
 Map<String, dynamic> _$NamedSpanPropertyToJson(NamedSpanProperty instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
 UndefinedSpanProperty _$UndefinedSpanPropertyFromJson(Map json) =>
-    UndefinedSpanProperty(
-      $type: json['type'] as String?,
-    );
+    UndefinedSpanProperty($type: json['type'] as String?);
 
 Map<String, dynamic> _$UndefinedSpanPropertyToJson(
-        UndefinedSpanProperty instance) =>
-    <String, dynamic>{
-      'type': instance.$type,
-    };
+  UndefinedSpanProperty instance,
+) => <String, dynamic>{'type': instance.$type};
 
 DefinedParagraphProperty _$DefinedParagraphPropertyFromJson(Map json) =>
     DefinedParagraphProperty(
       span: json['span'] == null
           ? const DefinedSpanProperty()
           : DefinedSpanProperty.fromJson(
-              Map<String, dynamic>.from(json['span'] as Map)),
-      alignment: $enumDecodeNullable(
-              _$HorizontalAlignmentEnumMap, json['alignment']) ??
+              Map<String, dynamic>.from(json['span'] as Map),
+            ),
+      alignment:
+          $enumDecodeNullable(
+            _$HorizontalAlignmentEnumMap,
+            json['alignment'],
+          ) ??
           HorizontalAlignment.left,
       $type: json['type'] as String?,
     );
 
 Map<String, dynamic> _$DefinedParagraphPropertyToJson(
-        DefinedParagraphProperty instance) =>
-    <String, dynamic>{
-      'span': instance.span.toJson(),
-      'alignment': _$HorizontalAlignmentEnumMap[instance.alignment]!,
-      'type': instance.$type,
-    };
+  DefinedParagraphProperty instance,
+) => <String, dynamic>{
+  'span': instance.span.toJson(),
+  'alignment': _$HorizontalAlignmentEnumMap[instance.alignment]!,
+  'type': instance.$type,
+};
 
 const _$HorizontalAlignmentEnumMap = {
   HorizontalAlignment.left: 'left',
@@ -124,28 +131,21 @@ NamedParagraphProperty _$NamedParagraphPropertyFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$NamedParagraphPropertyToJson(
-        NamedParagraphProperty instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+  NamedParagraphProperty instance,
+) => <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
 UndefinedParagraphProperty _$UndefinedParagraphPropertyFromJson(Map json) =>
-    UndefinedParagraphProperty(
-      $type: json['type'] as String?,
-    );
+    UndefinedParagraphProperty($type: json['type'] as String?);
 
 Map<String, dynamic> _$UndefinedParagraphPropertyToJson(
-        UndefinedParagraphProperty instance) =>
-    <String, dynamic>{
-      'type': instance.$type,
-    };
+  UndefinedParagraphProperty instance,
+) => <String, dynamic>{'type': instance.$type};
 
 _AreaProperty _$AreaPropertyFromJson(Map json) => _AreaProperty(
-      alignment:
-          $enumDecodeNullable(_$VerticalAlignmentEnumMap, json['alignment']) ??
-              VerticalAlignment.top,
-    );
+  alignment:
+      $enumDecodeNullable(_$VerticalAlignmentEnumMap, json['alignment']) ??
+      VerticalAlignment.top,
+);
 
 Map<String, dynamic> _$AreaPropertyToJson(_AreaProperty instance) =>
     <String, dynamic>{
@@ -159,29 +159,31 @@ const _$VerticalAlignmentEnumMap = {
 };
 
 _TextSpan _$TextSpanFromJson(Map json) => _TextSpan(
-      text: json['text'] as String? ?? '',
-      property: json['property'] == null
-          ? const SpanProperty.undefined()
-          : SpanProperty.fromJson(
-              Map<String, dynamic>.from(json['property'] as Map)),
-    );
+  text: json['text'] as String? ?? '',
+  property: json['property'] == null
+      ? const SpanProperty.undefined()
+      : SpanProperty.fromJson(
+          Map<String, dynamic>.from(json['property'] as Map),
+        ),
+);
 
 Map<String, dynamic> _$TextSpanToJson(_TextSpan instance) => <String, dynamic>{
-      'text': instance.text,
-      'property': instance.property.toJson(),
-    };
+  'text': instance.text,
+  'property': instance.property.toJson(),
+};
 
 _TextParagraph _$TextParagraphFromJson(Map json) => _TextParagraph(
-      property: json['property'] == null
-          ? const ParagraphProperty.undefined()
-          : ParagraphProperty.fromJson(
-              Map<String, dynamic>.from(json['property'] as Map)),
-      textSpans: (json['textSpans'] as List<dynamic>?)
-              ?.map(
-                  (e) => TextSpan.fromJson(Map<String, dynamic>.from(e as Map)))
-              .toList() ??
-          const [],
-    );
+  property: json['property'] == null
+      ? const ParagraphProperty.undefined()
+      : ParagraphProperty.fromJson(
+          Map<String, dynamic>.from(json['property'] as Map),
+        ),
+  textSpans:
+      (json['textSpans'] as List<dynamic>?)
+          ?.map((e) => TextSpan.fromJson(Map<String, dynamic>.from(e as Map)))
+          .toList() ??
+      const [],
+);
 
 Map<String, dynamic> _$TextParagraphToJson(_TextParagraph instance) =>
     <String, dynamic>{
@@ -190,40 +192,48 @@ Map<String, dynamic> _$TextParagraphToJson(_TextParagraph instance) =>
     };
 
 _TextArea _$TextAreaFromJson(Map json) => _TextArea(
-      areaProperty: json['areaProperty'] == null
-          ? const AreaProperty()
-          : AreaProperty.fromJson(
-              Map<String, dynamic>.from(json['areaProperty'] as Map)),
-      paragraph: TextParagraph.fromJson(
-          Map<String, dynamic>.from(json['paragraph'] as Map)),
-    );
+  areaProperty: json['areaProperty'] == null
+      ? const AreaProperty()
+      : AreaProperty.fromJson(
+          Map<String, dynamic>.from(json['areaProperty'] as Map),
+        ),
+  paragraph: TextParagraph.fromJson(
+    Map<String, dynamic>.from(json['paragraph'] as Map),
+  ),
+);
 
 Map<String, dynamic> _$TextAreaToJson(_TextArea instance) => <String, dynamic>{
-      'areaProperty': instance.areaProperty.toJson(),
-      'paragraph': instance.paragraph.toJson(),
-    };
+  'areaProperty': instance.areaProperty.toJson(),
+  'paragraph': instance.paragraph.toJson(),
+};
 
 _TextStyleSheet _$TextStyleSheetFromJson(Map json) => _TextStyleSheet(
-      spanProperties: (json['spanProperties'] as Map?)?.map(
-            (k, e) => MapEntry(
-                k as String,
-                DefinedSpanProperty.fromJson(
-                    Map<String, dynamic>.from(e as Map))),
-          ) ??
-          const {},
-      paragraphProperties: (json['paragraphProperties'] as Map?)?.map(
-            (k, e) => MapEntry(
-                k as String,
-                DefinedParagraphProperty.fromJson(
-                    Map<String, dynamic>.from(e as Map))),
-          ) ??
-          const {},
-    );
+  spanProperties:
+      (json['spanProperties'] as Map?)?.map(
+        (k, e) => MapEntry(
+          k as String,
+          DefinedSpanProperty.fromJson(Map<String, dynamic>.from(e as Map)),
+        ),
+      ) ??
+      const {},
+  paragraphProperties:
+      (json['paragraphProperties'] as Map?)?.map(
+        (k, e) => MapEntry(
+          k as String,
+          DefinedParagraphProperty.fromJson(
+            Map<String, dynamic>.from(e as Map),
+          ),
+        ),
+      ) ??
+      const {},
+);
 
 Map<String, dynamic> _$TextStyleSheetToJson(_TextStyleSheet instance) =>
     <String, dynamic>{
-      'spanProperties':
-          instance.spanProperties.map((k, e) => MapEntry(k, e.toJson())),
-      'paragraphProperties':
-          instance.paragraphProperties.map((k, e) => MapEntry(k, e.toJson())),
+      'spanProperties': instance.spanProperties.map(
+        (k, e) => MapEntry(k, e.toJson()),
+      ),
+      'paragraphProperties': instance.paragraphProperties.map(
+        (k, e) => MapEntry(k, e.toJson()),
+      ),
     };
