@@ -15,7 +15,7 @@ enum PatternTemplate {
   plainDark,
   ruledDark,
   quadDark,
-  musicDark
+  musicDark,
 }
 
 extension PatternTemplateExtension on PatternTemplate {
@@ -24,23 +24,18 @@ extension PatternTemplateExtension on PatternTemplate {
       'templates/${name.replaceAllMapped(RegExp(r'([A-Z])'), (match) => '_${match.group(1)?.toLowerCase()}')}.png';
 
   bool get dark => [
-        PatternTemplate.plainDark,
-        PatternTemplate.ruledDark,
-        PatternTemplate.quadDark,
-        PatternTemplate.musicDark
-      ].contains(this);
+    PatternTemplate.plainDark,
+    PatternTemplate.ruledDark,
+    PatternTemplate.quadDark,
+    PatternTemplate.musicDark,
+  ].contains(this);
 
   PatternTexture create() {
     switch (this) {
       case PatternTemplate.plain:
-        return const PatternTexture(
-          boxColor: BasicColors.light,
-        );
+        return const PatternTexture(boxColor: BasicColors.light);
       case PatternTemplate.ruled:
-        return const PatternTexture(
-          boxColor: BasicColors.light,
-          boxHeight: 40,
-        );
+        return const PatternTexture(boxColor: BasicColors.light, boxHeight: 40);
       case PatternTemplate.quad:
         return const PatternTexture(
           boxColor: BasicColors.light,
@@ -49,15 +44,14 @@ extension PatternTemplateExtension on PatternTemplate {
         );
       case PatternTemplate.music:
         return const PatternTexture(
-            boxColor: BasicColors.light,
-            boxHeight: 40,
-            boxYColor: SRGBColor.black,
-            boxYSpace: 80,
-            boxYCount: 5);
-      case PatternTemplate.plainDark:
-        return const PatternTexture(
-          boxColor: BasicColors.dark,
+          boxColor: BasicColors.light,
+          boxHeight: 40,
+          boxYColor: SRGBColor.black,
+          boxYSpace: 80,
+          boxYCount: 5,
         );
+      case PatternTemplate.plainDark:
+        return const PatternTexture(boxColor: BasicColors.dark);
       case PatternTemplate.ruledDark:
         return const PatternTexture(boxColor: BasicColors.dark, boxHeight: 40);
       case PatternTemplate.quadDark:
@@ -68,29 +62,31 @@ extension PatternTemplateExtension on PatternTemplate {
         );
       case PatternTemplate.musicDark:
         return const PatternTexture(
-            boxColor: BasicColors.dark,
-            boxYColor: SRGBColor.white,
-            boxHeight: 40,
-            boxYSpace: 80,
-            boxYCount: 5);
+          boxColor: BasicColors.dark,
+          boxYColor: SRGBColor.white,
+          boxHeight: 40,
+          boxYSpace: 80,
+          boxYCount: 5,
+        );
     }
   }
 }
 
 @freezed
 sealed class SurfaceTexture with _$SurfaceTexture {
-  const factory SurfaceTexture.pattern(
-      {@Default(0) double boxWidth,
-      @Default(0) double boxHeight,
-      @Default(1) int boxXCount,
-      @Default(1) int boxYCount,
-      @Default(0) double boxXSpace,
-      @Default(0) double boxYSpace,
-      @Default(BasicColors.blue) @ColorJsonConverter() SRGBColor boxXColor,
-      @Default(BasicColors.red) @ColorJsonConverter() SRGBColor boxYColor,
-      @Default(SRGBColor.white) @ColorJsonConverter() SRGBColor boxColor,
-      @Default(0.5) double boxXStroke,
-      @Default(0.5) double boxYStroke}) = PatternTexture;
+  const factory SurfaceTexture.pattern({
+    @Default(0) double boxWidth,
+    @Default(0) double boxHeight,
+    @Default(1) int boxXCount,
+    @Default(1) int boxYCount,
+    @Default(0) double boxXSpace,
+    @Default(0) double boxYSpace,
+    @Default(BasicColors.blue) @ColorJsonConverter() SRGBColor boxXColor,
+    @Default(BasicColors.red) @ColorJsonConverter() SRGBColor boxYColor,
+    @Default(SRGBColor.white) @ColorJsonConverter() SRGBColor boxColor,
+    @Default(0.5) double boxXStroke,
+    @Default(0.5) double boxYStroke,
+  }) = PatternTexture;
 
   factory SurfaceTexture.fromJson(Map<String, dynamic> json) =>
       _$SurfaceTextureFromJson(json);
