@@ -23,9 +23,7 @@ sealed class DocumentEvent extends ReplayEvent with _$DocumentEvent {
   const factory DocumentEvent.pageAdded([int? index, DocumentPage? page]) =
       PageAdded;
 
-  const factory DocumentEvent.pageChanged(
-    String pageName,
-  ) = PageChanged;
+  const factory DocumentEvent.pageChanged(String pageName) = PageChanged;
 
   const factory DocumentEvent.pageReordered(String page, [int? newIndex]) =
       PageReordered;
@@ -44,17 +42,15 @@ sealed class DocumentEvent extends ReplayEvent with _$DocumentEvent {
   const factory DocumentEvent.utilitiesChanged(UtilitiesState state) =
       UtilitiesChanged;
 
-  const factory DocumentEvent.elementsCreated(
-    List<PadElement> elements,
-  ) = ElementsCreated;
+  const factory DocumentEvent.elementsCreated(List<PadElement> elements) =
+      ElementsCreated;
 
   const factory DocumentEvent.elementsChanged(
     Map<String, List<PadElement>> elements,
   ) = ElementsChanged;
 
-  const factory DocumentEvent.elementsRemoved(
-    List<String> elements,
-  ) = ElementsRemoved;
+  const factory DocumentEvent.elementsRemoved(List<String> elements) =
+      ElementsRemoved;
 
   const factory DocumentEvent.elementsArranged(
     Arrangement arrangement,
@@ -66,158 +62,119 @@ sealed class DocumentEvent extends ReplayEvent with _$DocumentEvent {
     String? description,
   }) = DocumentDescriptionChanged;
 
-  const factory DocumentEvent.documentSaved(
-      [@JsonKey(
-          fromJson: AssetLocationMapper.fromMap, toJson: _assetLocationToJson)
-      AssetLocation? location]) = DocumentSaved;
+  const factory DocumentEvent.documentSaved([
+    @JsonKey(
+      fromJson: AssetLocationMapper.fromMap,
+      toJson: _assetLocationToJson,
+    )
+    AssetLocation? location,
+  ]) = DocumentSaved;
 
-  const factory DocumentEvent.toolCreated(
-    Tool tool,
-  ) = ToolCreated;
+  const factory DocumentEvent.toolCreated(Tool tool) = ToolCreated;
 
-  const factory DocumentEvent.toolsChanged(
-    Map<int, Tool> tools,
-  ) = ToolsChanged;
+  const factory DocumentEvent.toolsChanged(Map<int, Tool> tools) = ToolsChanged;
 
-  const factory DocumentEvent.toolsRemoved(
-    List<int> tools,
-  ) = ToolsRemoved;
+  const factory DocumentEvent.toolsRemoved(List<int> tools) = ToolsRemoved;
 
-  const factory DocumentEvent.toolReordered(
-    int oldIndex,
-    int newIndex,
-  ) = ToolReordered;
+  const factory DocumentEvent.toolReordered(int oldIndex, int newIndex) =
+      ToolReordered;
 
   const factory DocumentEvent.documentBackgroundsChanged(
     List<Background> backgrounds,
   ) = DocumentBackgroundsChanged;
 
-  const factory DocumentEvent.waypointCreated(
-    Waypoint waypoint,
-  ) = WaypointCreated;
+  const factory DocumentEvent.waypointCreated(Waypoint waypoint) =
+      WaypointCreated;
 
-  const factory DocumentEvent.waypointChanged(
-    String? name,
-    Waypoint waypoint,
-  ) = WaypointChanged;
+  const factory DocumentEvent.waypointChanged(String? name, Waypoint waypoint) =
+      WaypointChanged;
 
-  const factory DocumentEvent.waypointRemoved(
-    String? name,
-  ) = WaypointRemoved;
+  const factory DocumentEvent.waypointRemoved(String? name) = WaypointRemoved;
 
-  const factory DocumentEvent.waypointReordered(
-    String? name,
-    int newIndex,
-  ) = WaypointReordered;
+  const factory DocumentEvent.waypointReordered(String? name, int newIndex) =
+      WaypointReordered;
 
   const factory DocumentEvent.collectionRenamed(
     String oldName,
     String newName,
   ) = CollectionRenamed;
 
-  const factory DocumentEvent.collectionElementsRemoved(
-    String name,
-  ) = CollectionElementsRemoved;
+  const factory DocumentEvent.collectionElementsRemoved(String name) =
+      CollectionElementsRemoved;
 
-  const factory DocumentEvent.currentCollectionChanged(
-    String name,
-  ) = CurrentCollectionChanged;
+  const factory DocumentEvent.currentCollectionChanged(String name) =
+      CurrentCollectionChanged;
 
-  const factory DocumentEvent.currentLayerChanged(
-    String name,
-  ) = CurrentLayerChanged;
+  const factory DocumentEvent.currentLayerChanged(String name) =
+      CurrentLayerChanged;
 
   const factory DocumentEvent.layerCreated({
     @Default('') String name,
     String? id,
   }) = LayerCreated;
 
-  const factory DocumentEvent.layerChanged(
-    String id, {
-    String? name,
-  }) = LayerChanged;
+  const factory DocumentEvent.layerChanged(String id, {String? name}) =
+      LayerChanged;
 
-  const factory DocumentEvent.layerRemoved(
-    String id,
-  ) = LayerRemoved;
+  const factory DocumentEvent.layerRemoved(String id) = LayerRemoved;
 
-  const factory DocumentEvent.layersMerged(List<String> layers,
-      [@Default(false) bool duplicate]) = LayersMerged;
+  const factory DocumentEvent.layersMerged(
+    List<String> layers, [
+    @Default(false) bool duplicate,
+  ]) = LayersMerged;
 
-  const factory DocumentEvent.layerOrderChanged(
-    String id,
-    int index,
-  ) = LayerOrderChanged;
+  const factory DocumentEvent.layerOrderChanged(String id, int index) =
+      LayerOrderChanged;
 
-  const factory DocumentEvent.layerVisibilityChanged(
-    String id,
-    bool visible,
-  ) = LayerVisibilityChanged;
+  const factory DocumentEvent.layerVisibilityChanged(String id, bool visible) =
+      LayerVisibilityChanged;
 
   const factory DocumentEvent.elementsCollectionChanged(
     List<String> elements,
     String collection,
   ) = ElementsCollectionChanged;
 
-  const factory DocumentEvent.areasCreated(
-    List<Area> areas,
-  ) = AreasCreated;
+  const factory DocumentEvent.areasCreated(List<Area> areas) = AreasCreated;
 
-  const factory DocumentEvent.areasRemoved(
-    List<String> areas,
-  ) = AreasRemoved;
+  const factory DocumentEvent.areasRemoved(List<String> areas) = AreasRemoved;
 
-  const factory DocumentEvent.areaChanged(
-    String name,
-    Area area,
-  ) = AreaChanged;
+  const factory DocumentEvent.areaChanged(String name, Area area) = AreaChanged;
 
-  const factory DocumentEvent.areaReordered(
-    String name,
-    int newIndex,
-  ) = AreaReordered;
+  const factory DocumentEvent.areaReordered(String name, int newIndex) =
+      AreaReordered;
 
-  const factory DocumentEvent.currentAreaChanged(
-    String name,
-  ) = CurrentAreaChanged;
+  const factory DocumentEvent.currentAreaChanged(String name) =
+      CurrentAreaChanged;
 
-  const factory DocumentEvent.exportPresetCreated(String name,
-      [@Default([]) List<AreaPreset> areas]) = ExportPresetCreated;
+  const factory DocumentEvent.exportPresetCreated(
+    String name, [
+    @Default([]) List<AreaPreset> areas,
+  ]) = ExportPresetCreated;
 
   const factory DocumentEvent.exportPresetUpdated(
     String name,
     List<AreaPreset> areas,
   ) = ExportPresetUpdated;
 
-  const factory DocumentEvent.exportPresetRemoved(
-    String name,
-  ) = ExportPresetRemoved;
+  const factory DocumentEvent.exportPresetRemoved(String name) =
+      ExportPresetRemoved;
 
-  const factory DocumentEvent.packAdded(
-    NoteData pack,
-  ) = PackAdded;
+  const factory DocumentEvent.packAdded(NoteData pack) = PackAdded;
 
-  const factory DocumentEvent.packUpdated(
-    String name,
-    NoteData pack,
-  ) = PackUpdated;
+  const factory DocumentEvent.packUpdated(String name, NoteData pack) =
+      PackUpdated;
 
-  const factory DocumentEvent.packRemoved(
-    String name,
-  ) = PackRemoved;
+  const factory DocumentEvent.packRemoved(String name) = PackRemoved;
 
-  const factory DocumentEvent.animationAdded(
-    AnimationTrack animation,
-  ) = AnimationAdded;
+  const factory DocumentEvent.animationAdded(AnimationTrack animation) =
+      AnimationAdded;
 
   const factory DocumentEvent.animationUpdated(
     String name,
     AnimationTrack animation,
   ) = AnimationUpdated;
 
-  const factory DocumentEvent.animationRemoved(
-    String name,
-  ) = AnimationRemoved;
+  const factory DocumentEvent.animationRemoved(String name) = AnimationRemoved;
 
   const factory DocumentEvent.presentationModeEntered(
     AnimationTrack track,
@@ -231,30 +188,33 @@ sealed class DocumentEvent extends ReplayEvent with _$DocumentEvent {
   const factory DocumentEvent.assetUpdated(String path, List<int> data) =
       AssetUpdated;
 
-  const factory DocumentEvent.elementsLayerConverted(List<String> elements,
-      [@Default('') String name]) = ElementsLayerConverted;
+  const factory DocumentEvent.elementsLayerConverted(
+    List<String> elements, [
+    @Default('') String name,
+  ]) = ElementsLayerConverted;
 
   const factory DocumentEvent.encryptionChanged(String? password) =
       EncryptionChanged;
 
   const factory DocumentEvent.documentRebuilt(
-      @Uint8ListJsonConverter() Uint8List data) = DocumentRebuilt;
+    @Uint8ListJsonConverter() Uint8List data,
+  ) = DocumentRebuilt;
 
   factory DocumentEvent.fromJson(Map<String, dynamic> json) =>
       _$DocumentEventFromJson(json);
 
   bool shouldSync() => switch (this) {
-        CurrentCollectionChanged _ => false,
-        DocumentSaved _ => false,
-        CurrentAreaChanged _ => false,
-        LayerVisibilityChanged _ => false,
-        UtilitiesChanged _ => false,
-        PresentationModeEntered _ => false,
-        PresentationModeExited _ => false,
-        PresentationTick _ => false,
-        DocumentRebuilt _ => false,
-        _ => true,
-      };
+    CurrentCollectionChanged _ => false,
+    DocumentSaved _ => false,
+    CurrentAreaChanged _ => false,
+    LayerVisibilityChanged _ => false,
+    UtilitiesChanged _ => false,
+    PresentationModeEntered _ => false,
+    PresentationModeExited _ => false,
+    PresentationTick _ => false,
+    DocumentRebuilt _ => false,
+    _ => true,
+  };
 
   bool isAllowed(DocumentPermission permission) {
     if (permission == DocumentPermission.read) return false;
