@@ -7,36 +7,30 @@ part of 'event.dart';
 // **************************************************************************
 
 PageAdded _$PageAddedFromJson(Map json) => PageAdded(
-      (json['index'] as num?)?.toInt(),
-      json['page'] == null
-          ? null
-          : DocumentPage.fromJson(
-              Map<String, dynamic>.from(json['page'] as Map)),
-      json['type'] as String?,
-    );
+  (json['index'] as num?)?.toInt(),
+  json['page'] == null
+      ? null
+      : DocumentPage.fromJson(Map<String, dynamic>.from(json['page'] as Map)),
+  json['type'] as String?,
+);
 
 Map<String, dynamic> _$PageAddedToJson(PageAdded instance) => <String, dynamic>{
-      'index': instance.index,
-      'page': instance.page?.toJson(),
-      'type': instance.$type,
-    };
+  'index': instance.index,
+  'page': instance.page?.toJson(),
+  'type': instance.$type,
+};
 
-PageChanged _$PageChangedFromJson(Map json) => PageChanged(
-      json['pageName'] as String,
-      $type: json['type'] as String?,
-    );
+PageChanged _$PageChangedFromJson(Map json) =>
+    PageChanged(json['pageName'] as String, $type: json['type'] as String?);
 
 Map<String, dynamic> _$PageChangedToJson(PageChanged instance) =>
-    <String, dynamic>{
-      'pageName': instance.pageName,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'pageName': instance.pageName, 'type': instance.$type};
 
 PageReordered _$PageReorderedFromJson(Map json) => PageReordered(
-      json['page'] as String,
-      (json['newIndex'] as num?)?.toInt(),
-      json['type'] as String?,
-    );
+  json['page'] as String,
+  (json['newIndex'] as num?)?.toInt(),
+  json['type'] as String?,
+);
 
 Map<String, dynamic> _$PageReorderedToJson(PageReordered instance) =>
     <String, dynamic>{
@@ -46,10 +40,10 @@ Map<String, dynamic> _$PageReorderedToJson(PageReordered instance) =>
     };
 
 PageRenamed _$PageRenamedFromJson(Map json) => PageRenamed(
-      json['oldName'] as String,
-      json['newName'] as String,
-      $type: json['type'] as String?,
-    );
+  json['oldName'] as String,
+  json['newName'] as String,
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$PageRenamedToJson(PageRenamed instance) =>
     <String, dynamic>{
@@ -58,21 +52,16 @@ Map<String, dynamic> _$PageRenamedToJson(PageRenamed instance) =>
       'type': instance.$type,
     };
 
-PageRemoved _$PageRemovedFromJson(Map json) => PageRemoved(
-      json['page'] as String,
-      $type: json['type'] as String?,
-    );
+PageRemoved _$PageRemovedFromJson(Map json) =>
+    PageRemoved(json['page'] as String, $type: json['type'] as String?);
 
 Map<String, dynamic> _$PageRemovedToJson(PageRemoved instance) =>
-    <String, dynamic>{
-      'page': instance.page,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'page': instance.page, 'type': instance.$type};
 
 ThumbnailCaptured _$ThumbnailCapturedFromJson(Map json) => ThumbnailCaptured(
-      const Uint8ListJsonConverter().fromJson(json['data'] as String),
-      $type: json['type'] as String?,
-    );
+  const Uint8ListJsonConverter().fromJson(json['data'] as String),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ThumbnailCapturedToJson(ThumbnailCaptured instance) =>
     <String, dynamic>{
@@ -81,33 +70,27 @@ Map<String, dynamic> _$ThumbnailCapturedToJson(ThumbnailCaptured instance) =>
     };
 
 ViewChanged _$ViewChangedFromJson(Map json) => ViewChanged(
-      ViewOption.fromJson(Map<String, dynamic>.from(json['view'] as Map)),
-      $type: json['type'] as String?,
-    );
+  ViewOption.fromJson(Map<String, dynamic>.from(json['view'] as Map)),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ViewChangedToJson(ViewChanged instance) =>
-    <String, dynamic>{
-      'view': instance.view.toJson(),
-      'type': instance.$type,
-    };
+    <String, dynamic>{'view': instance.view.toJson(), 'type': instance.$type};
 
 UtilitiesChanged _$UtilitiesChangedFromJson(Map json) => UtilitiesChanged(
-      UtilitiesState.fromJson(Map<String, dynamic>.from(json['state'] as Map)),
-      $type: json['type'] as String?,
-    );
+  UtilitiesState.fromJson(Map<String, dynamic>.from(json['state'] as Map)),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$UtilitiesChangedToJson(UtilitiesChanged instance) =>
-    <String, dynamic>{
-      'state': instance.state.toJson(),
-      'type': instance.$type,
-    };
+    <String, dynamic>{'state': instance.state.toJson(), 'type': instance.$type};
 
 ElementsCreated _$ElementsCreatedFromJson(Map json) => ElementsCreated(
-      (json['elements'] as List<dynamic>)
-          .map((e) => PadElement.fromJson(Map<String, dynamic>.from(e as Map)))
-          .toList(),
-      $type: json['type'] as String?,
-    );
+  (json['elements'] as List<dynamic>)
+      .map((e) => PadElement.fromJson(Map<String, dynamic>.from(e as Map)))
+      .toList(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ElementsCreatedToJson(ElementsCreated instance) =>
     <String, dynamic>{
@@ -116,40 +99,38 @@ Map<String, dynamic> _$ElementsCreatedToJson(ElementsCreated instance) =>
     };
 
 ElementsChanged _$ElementsChangedFromJson(Map json) => ElementsChanged(
-      (json['elements'] as Map).map(
-        (k, e) => MapEntry(
-            k as String,
-            (e as List<dynamic>)
-                .map((e) =>
-                    PadElement.fromJson(Map<String, dynamic>.from(e as Map)))
-                .toList()),
-      ),
-      $type: json['type'] as String?,
-    );
+  (json['elements'] as Map).map(
+    (k, e) => MapEntry(
+      k as String,
+      (e as List<dynamic>)
+          .map((e) => PadElement.fromJson(Map<String, dynamic>.from(e as Map)))
+          .toList(),
+    ),
+  ),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ElementsChangedToJson(ElementsChanged instance) =>
     <String, dynamic>{
-      'elements': instance.elements
-          .map((k, e) => MapEntry(k, e.map((e) => e.toJson()).toList())),
+      'elements': instance.elements.map(
+        (k, e) => MapEntry(k, e.map((e) => e.toJson()).toList()),
+      ),
       'type': instance.$type,
     };
 
 ElementsRemoved _$ElementsRemovedFromJson(Map json) => ElementsRemoved(
-      (json['elements'] as List<dynamic>).map((e) => e as String).toList(),
-      $type: json['type'] as String?,
-    );
+  (json['elements'] as List<dynamic>).map((e) => e as String).toList(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ElementsRemovedToJson(ElementsRemoved instance) =>
-    <String, dynamic>{
-      'elements': instance.elements,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'elements': instance.elements, 'type': instance.$type};
 
 ElementsArranged _$ElementsArrangedFromJson(Map json) => ElementsArranged(
-      $enumDecode(_$ArrangementEnumMap, json['arrangement']),
-      (json['elements'] as List<dynamic>).map((e) => e as String).toList(),
-      $type: json['type'] as String?,
-    );
+  $enumDecode(_$ArrangementEnumMap, json['arrangement']),
+  (json['elements'] as List<dynamic>).map((e) => e as String).toList(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ElementsArrangedToJson(ElementsArranged instance) =>
     <String, dynamic>{
@@ -173,17 +154,17 @@ DocumentDescriptionChanged _$DocumentDescriptionChangedFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$DocumentDescriptionChangedToJson(
-        DocumentDescriptionChanged instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'description': instance.description,
-      'type': instance.$type,
-    };
+  DocumentDescriptionChanged instance,
+) => <String, dynamic>{
+  'name': instance.name,
+  'description': instance.description,
+  'type': instance.$type,
+};
 
 DocumentSaved _$DocumentSavedFromJson(Map json) => DocumentSaved(
-      AssetLocationMapper.fromMap(json['location'] as Map<String, dynamic>),
-      json['type'] as String?,
-    );
+  AssetLocationMapper.fromMap(json['location'] as Map<String, dynamic>),
+  json['type'] as String?,
+);
 
 Map<String, dynamic> _$DocumentSavedToJson(DocumentSaved instance) =>
     <String, dynamic>{
@@ -192,23 +173,22 @@ Map<String, dynamic> _$DocumentSavedToJson(DocumentSaved instance) =>
     };
 
 ToolCreated _$ToolCreatedFromJson(Map json) => ToolCreated(
-      Tool.fromJson(Map<String, dynamic>.from(json['tool'] as Map)),
-      $type: json['type'] as String?,
-    );
+  Tool.fromJson(Map<String, dynamic>.from(json['tool'] as Map)),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ToolCreatedToJson(ToolCreated instance) =>
-    <String, dynamic>{
-      'tool': instance.tool.toJson(),
-      'type': instance.$type,
-    };
+    <String, dynamic>{'tool': instance.tool.toJson(), 'type': instance.$type};
 
 ToolsChanged _$ToolsChangedFromJson(Map json) => ToolsChanged(
-      (json['tools'] as Map).map(
-        (k, e) => MapEntry(int.parse(k as String),
-            Tool.fromJson(Map<String, dynamic>.from(e as Map))),
-      ),
-      $type: json['type'] as String?,
-    );
+  (json['tools'] as Map).map(
+    (k, e) => MapEntry(
+      int.parse(k as String),
+      Tool.fromJson(Map<String, dynamic>.from(e as Map)),
+    ),
+  ),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ToolsChangedToJson(ToolsChanged instance) =>
     <String, dynamic>{
@@ -217,21 +197,18 @@ Map<String, dynamic> _$ToolsChangedToJson(ToolsChanged instance) =>
     };
 
 ToolsRemoved _$ToolsRemovedFromJson(Map json) => ToolsRemoved(
-      (json['tools'] as List<dynamic>).map((e) => (e as num).toInt()).toList(),
-      $type: json['type'] as String?,
-    );
+  (json['tools'] as List<dynamic>).map((e) => (e as num).toInt()).toList(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ToolsRemovedToJson(ToolsRemoved instance) =>
-    <String, dynamic>{
-      'tools': instance.tools,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'tools': instance.tools, 'type': instance.$type};
 
 ToolReordered _$ToolReorderedFromJson(Map json) => ToolReordered(
-      (json['oldIndex'] as num).toInt(),
-      (json['newIndex'] as num).toInt(),
-      $type: json['type'] as String?,
-    );
+  (json['oldIndex'] as num).toInt(),
+  (json['newIndex'] as num).toInt(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$ToolReorderedToJson(ToolReordered instance) =>
     <String, dynamic>{
@@ -249,16 +226,16 @@ DocumentBackgroundsChanged _$DocumentBackgroundsChangedFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$DocumentBackgroundsChangedToJson(
-        DocumentBackgroundsChanged instance) =>
-    <String, dynamic>{
-      'backgrounds': instance.backgrounds.map((e) => e.toJson()).toList(),
-      'type': instance.$type,
-    };
+  DocumentBackgroundsChanged instance,
+) => <String, dynamic>{
+  'backgrounds': instance.backgrounds.map((e) => e.toJson()).toList(),
+  'type': instance.$type,
+};
 
 WaypointCreated _$WaypointCreatedFromJson(Map json) => WaypointCreated(
-      Waypoint.fromJson(Map<String, dynamic>.from(json['waypoint'] as Map)),
-      $type: json['type'] as String?,
-    );
+  Waypoint.fromJson(Map<String, dynamic>.from(json['waypoint'] as Map)),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$WaypointCreatedToJson(WaypointCreated instance) =>
     <String, dynamic>{
@@ -267,10 +244,10 @@ Map<String, dynamic> _$WaypointCreatedToJson(WaypointCreated instance) =>
     };
 
 WaypointChanged _$WaypointChangedFromJson(Map json) => WaypointChanged(
-      json['name'] as String?,
-      Waypoint.fromJson(Map<String, dynamic>.from(json['waypoint'] as Map)),
-      $type: json['type'] as String?,
-    );
+  json['name'] as String?,
+  Waypoint.fromJson(Map<String, dynamic>.from(json['waypoint'] as Map)),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$WaypointChangedToJson(WaypointChanged instance) =>
     <String, dynamic>{
@@ -279,22 +256,17 @@ Map<String, dynamic> _$WaypointChangedToJson(WaypointChanged instance) =>
       'type': instance.$type,
     };
 
-WaypointRemoved _$WaypointRemovedFromJson(Map json) => WaypointRemoved(
-      json['name'] as String?,
-      $type: json['type'] as String?,
-    );
+WaypointRemoved _$WaypointRemovedFromJson(Map json) =>
+    WaypointRemoved(json['name'] as String?, $type: json['type'] as String?);
 
 Map<String, dynamic> _$WaypointRemovedToJson(WaypointRemoved instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
 WaypointReordered _$WaypointReorderedFromJson(Map json) => WaypointReordered(
-      json['name'] as String?,
-      (json['newIndex'] as num).toInt(),
-      $type: json['type'] as String?,
-    );
+  json['name'] as String?,
+  (json['newIndex'] as num).toInt(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$WaypointReorderedToJson(WaypointReordered instance) =>
     <String, dynamic>{
@@ -304,10 +276,10 @@ Map<String, dynamic> _$WaypointReorderedToJson(WaypointReordered instance) =>
     };
 
 CollectionRenamed _$CollectionRenamedFromJson(Map json) => CollectionRenamed(
-      json['oldName'] as String,
-      json['newName'] as String,
-      $type: json['type'] as String?,
-    );
+  json['oldName'] as String,
+  json['newName'] as String,
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$CollectionRenamedToJson(CollectionRenamed instance) =>
     <String, dynamic>{
@@ -323,11 +295,8 @@ CollectionElementsRemoved _$CollectionElementsRemovedFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$CollectionElementsRemovedToJson(
-        CollectionElementsRemoved instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+  CollectionElementsRemoved instance,
+) => <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
 CurrentCollectionChanged _$CurrentCollectionChangedFromJson(Map json) =>
     CurrentCollectionChanged(
@@ -336,30 +305,21 @@ CurrentCollectionChanged _$CurrentCollectionChangedFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$CurrentCollectionChangedToJson(
-        CurrentCollectionChanged instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+  CurrentCollectionChanged instance,
+) => <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
 CurrentLayerChanged _$CurrentLayerChangedFromJson(Map json) =>
-    CurrentLayerChanged(
-      json['name'] as String,
-      $type: json['type'] as String?,
-    );
+    CurrentLayerChanged(json['name'] as String, $type: json['type'] as String?);
 
 Map<String, dynamic> _$CurrentLayerChangedToJson(
-        CurrentLayerChanged instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+  CurrentLayerChanged instance,
+) => <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
 LayerCreated _$LayerCreatedFromJson(Map json) => LayerCreated(
-      name: json['name'] as String? ?? '',
-      id: json['id'] as String?,
-      $type: json['type'] as String?,
-    );
+  name: json['name'] as String? ?? '',
+  id: json['id'] as String?,
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$LayerCreatedToJson(LayerCreated instance) =>
     <String, dynamic>{
@@ -369,10 +329,10 @@ Map<String, dynamic> _$LayerCreatedToJson(LayerCreated instance) =>
     };
 
 LayerChanged _$LayerChangedFromJson(Map json) => LayerChanged(
-      json['id'] as String,
-      name: json['name'] as String?,
-      $type: json['type'] as String?,
-    );
+  json['id'] as String,
+  name: json['name'] as String?,
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$LayerChangedToJson(LayerChanged instance) =>
     <String, dynamic>{
@@ -381,22 +341,17 @@ Map<String, dynamic> _$LayerChangedToJson(LayerChanged instance) =>
       'type': instance.$type,
     };
 
-LayerRemoved _$LayerRemovedFromJson(Map json) => LayerRemoved(
-      json['id'] as String,
-      $type: json['type'] as String?,
-    );
+LayerRemoved _$LayerRemovedFromJson(Map json) =>
+    LayerRemoved(json['id'] as String, $type: json['type'] as String?);
 
 Map<String, dynamic> _$LayerRemovedToJson(LayerRemoved instance) =>
-    <String, dynamic>{
-      'id': instance.id,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'id': instance.id, 'type': instance.$type};
 
 LayersMerged _$LayersMergedFromJson(Map json) => LayersMerged(
-      (json['layers'] as List<dynamic>).map((e) => e as String).toList(),
-      json['duplicate'] as bool? ?? false,
-      json['type'] as String?,
-    );
+  (json['layers'] as List<dynamic>).map((e) => e as String).toList(),
+  json['duplicate'] as bool? ?? false,
+  json['type'] as String?,
+);
 
 Map<String, dynamic> _$LayersMergedToJson(LayersMerged instance) =>
     <String, dynamic>{
@@ -406,10 +361,10 @@ Map<String, dynamic> _$LayersMergedToJson(LayersMerged instance) =>
     };
 
 LayerOrderChanged _$LayerOrderChangedFromJson(Map json) => LayerOrderChanged(
-      json['id'] as String,
-      (json['index'] as num).toInt(),
-      $type: json['type'] as String?,
-    );
+  json['id'] as String,
+  (json['index'] as num).toInt(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$LayerOrderChangedToJson(LayerOrderChanged instance) =>
     <String, dynamic>{
@@ -426,12 +381,12 @@ LayerVisibilityChanged _$LayerVisibilityChangedFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$LayerVisibilityChangedToJson(
-        LayerVisibilityChanged instance) =>
-    <String, dynamic>{
-      'id': instance.id,
-      'visible': instance.visible,
-      'type': instance.$type,
-    };
+  LayerVisibilityChanged instance,
+) => <String, dynamic>{
+  'id': instance.id,
+  'visible': instance.visible,
+  'type': instance.$type,
+};
 
 ElementsCollectionChanged _$ElementsCollectionChangedFromJson(Map json) =>
     ElementsCollectionChanged(
@@ -441,19 +396,19 @@ ElementsCollectionChanged _$ElementsCollectionChangedFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$ElementsCollectionChangedToJson(
-        ElementsCollectionChanged instance) =>
-    <String, dynamic>{
-      'elements': instance.elements,
-      'collection': instance.collection,
-      'type': instance.$type,
-    };
+  ElementsCollectionChanged instance,
+) => <String, dynamic>{
+  'elements': instance.elements,
+  'collection': instance.collection,
+  'type': instance.$type,
+};
 
 AreasCreated _$AreasCreatedFromJson(Map json) => AreasCreated(
-      (json['areas'] as List<dynamic>)
-          .map((e) => Area.fromJson(Map<String, dynamic>.from(e as Map)))
-          .toList(),
-      $type: json['type'] as String?,
-    );
+  (json['areas'] as List<dynamic>)
+      .map((e) => Area.fromJson(Map<String, dynamic>.from(e as Map)))
+      .toList(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$AreasCreatedToJson(AreasCreated instance) =>
     <String, dynamic>{
@@ -462,21 +417,18 @@ Map<String, dynamic> _$AreasCreatedToJson(AreasCreated instance) =>
     };
 
 AreasRemoved _$AreasRemovedFromJson(Map json) => AreasRemoved(
-      (json['areas'] as List<dynamic>).map((e) => e as String).toList(),
-      $type: json['type'] as String?,
-    );
+  (json['areas'] as List<dynamic>).map((e) => e as String).toList(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$AreasRemovedToJson(AreasRemoved instance) =>
-    <String, dynamic>{
-      'areas': instance.areas,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'areas': instance.areas, 'type': instance.$type};
 
 AreaChanged _$AreaChangedFromJson(Map json) => AreaChanged(
-      json['name'] as String,
-      Area.fromJson(Map<String, dynamic>.from(json['area'] as Map)),
-      $type: json['type'] as String?,
-    );
+  json['name'] as String,
+  Area.fromJson(Map<String, dynamic>.from(json['area'] as Map)),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$AreaChangedToJson(AreaChanged instance) =>
     <String, dynamic>{
@@ -486,10 +438,10 @@ Map<String, dynamic> _$AreaChangedToJson(AreaChanged instance) =>
     };
 
 AreaReordered _$AreaReorderedFromJson(Map json) => AreaReordered(
-      json['name'] as String,
-      (json['newIndex'] as num).toInt(),
-      $type: json['type'] as String?,
-    );
+  json['name'] as String,
+  (json['newIndex'] as num).toInt(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$AreaReorderedToJson(AreaReordered instance) =>
     <String, dynamic>{
@@ -498,35 +450,31 @@ Map<String, dynamic> _$AreaReorderedToJson(AreaReordered instance) =>
       'type': instance.$type,
     };
 
-CurrentAreaChanged _$CurrentAreaChangedFromJson(Map json) => CurrentAreaChanged(
-      json['name'] as String,
-      $type: json['type'] as String?,
-    );
+CurrentAreaChanged _$CurrentAreaChangedFromJson(Map json) =>
+    CurrentAreaChanged(json['name'] as String, $type: json['type'] as String?);
 
 Map<String, dynamic> _$CurrentAreaChangedToJson(CurrentAreaChanged instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
 ExportPresetCreated _$ExportPresetCreatedFromJson(Map json) =>
     ExportPresetCreated(
       json['name'] as String,
       (json['areas'] as List<dynamic>?)
-              ?.map((e) =>
-                  AreaPreset.fromJson(Map<String, dynamic>.from(e as Map)))
+              ?.map(
+                (e) => AreaPreset.fromJson(Map<String, dynamic>.from(e as Map)),
+              )
               .toList() ??
           const [],
       json['type'] as String?,
     );
 
 Map<String, dynamic> _$ExportPresetCreatedToJson(
-        ExportPresetCreated instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'areas': instance.areas.map((e) => e.toJson()).toList(),
-      'type': instance.$type,
-    };
+  ExportPresetCreated instance,
+) => <String, dynamic>{
+  'name': instance.name,
+  'areas': instance.areas.map((e) => e.toJson()).toList(),
+  'type': instance.$type,
+};
 
 ExportPresetUpdated _$ExportPresetUpdatedFromJson(Map json) =>
     ExportPresetUpdated(
@@ -538,41 +486,33 @@ ExportPresetUpdated _$ExportPresetUpdatedFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$ExportPresetUpdatedToJson(
-        ExportPresetUpdated instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'areas': instance.areas.map((e) => e.toJson()).toList(),
-      'type': instance.$type,
-    };
+  ExportPresetUpdated instance,
+) => <String, dynamic>{
+  'name': instance.name,
+  'areas': instance.areas.map((e) => e.toJson()).toList(),
+  'type': instance.$type,
+};
 
 ExportPresetRemoved _$ExportPresetRemovedFromJson(Map json) =>
-    ExportPresetRemoved(
-      json['name'] as String,
-      $type: json['type'] as String?,
-    );
+    ExportPresetRemoved(json['name'] as String, $type: json['type'] as String?);
 
 Map<String, dynamic> _$ExportPresetRemovedToJson(
-        ExportPresetRemoved instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+  ExportPresetRemoved instance,
+) => <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
-PackAdded _$PackAddedFromJson(Map json) => PackAdded(
-      NoteData.fromJson(json['pack']),
-      $type: json['type'] as String?,
-    );
+PackAdded _$PackAddedFromJson(Map json) =>
+    PackAdded(NoteData.fromJson(json['pack']), $type: json['type'] as String?);
 
 Map<String, dynamic> _$PackAddedToJson(PackAdded instance) => <String, dynamic>{
-      'pack': instance.pack.toJson(),
-      'type': instance.$type,
-    };
+  'pack': instance.pack.toJson(),
+  'type': instance.$type,
+};
 
 PackUpdated _$PackUpdatedFromJson(Map json) => PackUpdated(
-      json['name'] as String,
-      NoteData.fromJson(json['pack']),
-      $type: json['type'] as String?,
-    );
+  json['name'] as String,
+  NoteData.fromJson(json['pack']),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$PackUpdatedToJson(PackUpdated instance) =>
     <String, dynamic>{
@@ -581,22 +521,16 @@ Map<String, dynamic> _$PackUpdatedToJson(PackUpdated instance) =>
       'type': instance.$type,
     };
 
-PackRemoved _$PackRemovedFromJson(Map json) => PackRemoved(
-      json['name'] as String,
-      $type: json['type'] as String?,
-    );
+PackRemoved _$PackRemovedFromJson(Map json) =>
+    PackRemoved(json['name'] as String, $type: json['type'] as String?);
 
 Map<String, dynamic> _$PackRemovedToJson(PackRemoved instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
 AnimationAdded _$AnimationAddedFromJson(Map json) => AnimationAdded(
-      AnimationTrack.fromJson(
-          Map<String, dynamic>.from(json['animation'] as Map)),
-      $type: json['type'] as String?,
-    );
+  AnimationTrack.fromJson(Map<String, dynamic>.from(json['animation'] as Map)),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$AnimationAddedToJson(AnimationAdded instance) =>
     <String, dynamic>{
@@ -605,11 +539,10 @@ Map<String, dynamic> _$AnimationAddedToJson(AnimationAdded instance) =>
     };
 
 AnimationUpdated _$AnimationUpdatedFromJson(Map json) => AnimationUpdated(
-      json['name'] as String,
-      AnimationTrack.fromJson(
-          Map<String, dynamic>.from(json['animation'] as Map)),
-      $type: json['type'] as String?,
-    );
+  json['name'] as String,
+  AnimationTrack.fromJson(Map<String, dynamic>.from(json['animation'] as Map)),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$AnimationUpdatedToJson(AnimationUpdated instance) =>
     <String, dynamic>{
@@ -618,16 +551,11 @@ Map<String, dynamic> _$AnimationUpdatedToJson(AnimationUpdated instance) =>
       'type': instance.$type,
     };
 
-AnimationRemoved _$AnimationRemovedFromJson(Map json) => AnimationRemoved(
-      json['name'] as String,
-      $type: json['type'] as String?,
-    );
+AnimationRemoved _$AnimationRemovedFromJson(Map json) =>
+    AnimationRemoved(json['name'] as String, $type: json['type'] as String?);
 
 Map<String, dynamic> _$AnimationRemovedToJson(AnimationRemoved instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'name': instance.name, 'type': instance.$type};
 
 PresentationModeEntered _$PresentationModeEnteredFromJson(Map json) =>
     PresentationModeEntered(
@@ -637,40 +565,33 @@ PresentationModeEntered _$PresentationModeEnteredFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$PresentationModeEnteredToJson(
-        PresentationModeEntered instance) =>
-    <String, dynamic>{
-      'track': instance.track.toJson(),
-      'fullScreen': instance.fullScreen,
-      'type': instance.$type,
-    };
+  PresentationModeEntered instance,
+) => <String, dynamic>{
+  'track': instance.track.toJson(),
+  'fullScreen': instance.fullScreen,
+  'type': instance.$type,
+};
 
 PresentationModeExited _$PresentationModeExitedFromJson(Map json) =>
-    PresentationModeExited(
-      $type: json['type'] as String?,
-    );
+    PresentationModeExited($type: json['type'] as String?);
 
 Map<String, dynamic> _$PresentationModeExitedToJson(
-        PresentationModeExited instance) =>
-    <String, dynamic>{
-      'type': instance.$type,
-    };
+  PresentationModeExited instance,
+) => <String, dynamic>{'type': instance.$type};
 
 PresentationTick _$PresentationTickFromJson(Map json) => PresentationTick(
-      (json['tick'] as num).toInt(),
-      $type: json['type'] as String?,
-    );
+  (json['tick'] as num).toInt(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$PresentationTickToJson(PresentationTick instance) =>
-    <String, dynamic>{
-      'tick': instance.tick,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'tick': instance.tick, 'type': instance.$type};
 
 AssetUpdated _$AssetUpdatedFromJson(Map json) => AssetUpdated(
-      json['path'] as String,
-      (json['data'] as List<dynamic>).map((e) => (e as num).toInt()).toList(),
-      $type: json['type'] as String?,
-    );
+  json['path'] as String,
+  (json['data'] as List<dynamic>).map((e) => (e as num).toInt()).toList(),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$AssetUpdatedToJson(AssetUpdated instance) =>
     <String, dynamic>{
@@ -687,28 +608,25 @@ ElementsLayerConverted _$ElementsLayerConvertedFromJson(Map json) =>
     );
 
 Map<String, dynamic> _$ElementsLayerConvertedToJson(
-        ElementsLayerConverted instance) =>
-    <String, dynamic>{
-      'elements': instance.elements,
-      'name': instance.name,
-      'type': instance.$type,
-    };
+  ElementsLayerConverted instance,
+) => <String, dynamic>{
+  'elements': instance.elements,
+  'name': instance.name,
+  'type': instance.$type,
+};
 
 EncryptionChanged _$EncryptionChangedFromJson(Map json) => EncryptionChanged(
-      json['password'] as String?,
-      $type: json['type'] as String?,
-    );
+  json['password'] as String?,
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$EncryptionChangedToJson(EncryptionChanged instance) =>
-    <String, dynamic>{
-      'password': instance.password,
-      'type': instance.$type,
-    };
+    <String, dynamic>{'password': instance.password, 'type': instance.$type};
 
 DocumentRebuilt _$DocumentRebuiltFromJson(Map json) => DocumentRebuilt(
-      const Uint8ListJsonConverter().fromJson(json['data'] as String),
-      $type: json['type'] as String?,
-    );
+  const Uint8ListJsonConverter().fromJson(json['data'] as String),
+  $type: json['type'] as String?,
+);
 
 Map<String, dynamic> _$DocumentRebuiltToJson(DocumentRebuilt instance) =>
     <String, dynamic>{
